==> 1. get fund_data from api.ipynb <==

from iFinDPy import *
from datetime import datetime
import pandas as pd
import time as _time
import json
from threading import Thread,Lock,Semaphore
import requests


THS_iFinDLogin("xnyf038","856030")

def datapool_hd():
    data_fund=THS_BD('000008.OF,000042.OF,000051.OF,000059.OF,000082.OF,000176.OF,000248.OF,000309.OF,000311.OF,000312.OF,000313.OF,000326.OF,000368.OF,000373.OF,000376.OF,000409.OF,000411.OF,000418.OF,000457.OF,000471.OF,000478.OF,000512.OF,000513.OF,000524.OF,000549.OF,000577.OF,000586.OF,000592.OF,000594.OF,000596.OF,000613.OF,000628.OF,000656.OF,000688.OF,000696.OF,000697.OF,000711.OF,000729.OF,000746.OF,000751.OF,000756.OF,000761.OF,000778.OF,000780.OF,000793.OF,000803.OF,000826.OF,000827.OF,000828.OF,000831.OF,000835.OF,000854.OF,000866.OF,000867.OF,000884.OF,000893.OF,000913.OF,000916.OF,000925.OF,000942.OF,000948.OF,000950.OF,000955.OF,000960.OF,000961.OF,000962.OF,000968.OF,000971.OF,000974.OF,000975.OF,000978.OF,000979.OF,000985.OF,000991.OF,000996.OF,001008.OF,001009.OF,001015.OF,001016.OF,001027.OF,001028.OF,001036.OF,001039.OF,001040.OF,001042.OF,001043.OF,001044.OF,001047.OF,001048.OF,001050.OF,001051.OF,001052.OF,001054.OF,001064.OF,001070.OF,001072.OF,001097.OF,001104.OF,001105.OF,001113.OF,001126.OF,001133.OF,001149.OF,001158.OF,001162.OF,001163.OF,001166.OF,001167.OF,001171.OF,001178.OF,001180.OF,001186.OF,001188.OF,001193.OF,001195.OF,001208.OF,001214.OF,001223.OF,001230.OF,001236.OF,001237.OF,001241.OF,001242.OF,001243.OF,001245.OF,001277.OF,001291.OF,001313.OF,001319.OF,001344.OF,001351.OF,001361.OF,001396.OF,001397.OF,001404.OF,001409.OF,001410.OF,001416.OF,001420.OF,001421.OF,001426.OF,001455.OF,001469.OF,001473.OF,001476.OF,001482.OF,001490.OF,001496.OF,001520.OF,001521.OF,001528.OF,001539.OF,001541.OF,001542.OF,001548.OF,001549.OF,001550.OF,001551.OF,001552.OF,001553.OF,001556.OF,001557.OF,001576.OF,001577.OF,001579.OF,001583.OF,001588.OF,001589.OF,001592.OF,001593.OF,001594.OF,001595.OF,001605.OF,001616.OF,001617.OF,001618.OF,001626.OF,001628.OF,001629.OF,001630.OF,001631.OF,001632.OF,001637.OF,001638.OF,001643.OF,001644.OF,001645.OF,001651.OF,001672.OF,001677.OF,001678.OF,001685.OF,001692.OF,001703.OF,001705.OF,001707.OF,001714.OF,001717.OF,001718.OF,001719.OF,001725.OF,001726.OF,001733.OF,001736.OF,001749.OF,001764.OF,001766.OF,001781.OF,001790.OF,001825.OF,001849.OF,001877.OF,001878.OF,001879.OF,001899.OF,001915.OF,001917.OF,001938.OF,001956.OF,001974.OF,001975.OF,002063.OF,002076.OF,002121.OF,002168.OF,002199.OF,002210.OF,002229.OF,002236.OF,002289.OF,002300.OF,002310.OF,002311.OF,002315.OF,002316.OF,002332.OF,002333.OF,002334.OF,002335.OF,002385.OF,002387.OF,002510.OF,002556.OF,002595.OF,002621.OF,002656.OF,002670.OF,002671.OF,002697.OF,002861.OF,002871.OF,002872.OF,002900.OF,002903.OF,002906.OF,002907.OF,002952.OF,002974.OF,002977.OF,002978.OF,002979.OF,002980.OF,002982.OF,002984.OF,002987.OF,003015.OF,003016.OF,003017.OF,003053.OF,003054.OF,003069.OF,003070.OF,003145.OF,003184.OF,003190.OF,003191.OF,003194.OF,003230.OF,003231.OF,003232.OF,003233.OF,003298.OF,003299.OF,003318.OF,003359.OF,003366.OF,003416.OF,003475.OF,003548.OF,003578.OF,003579.OF,003624.OF,003625.OF,003634.OF,003646.OF,003647.OF,003745.OF,003760.OF,003761.OF,003765.OF,003766.OF,003834.OF,003853.OF,003865.OF,003876.OF,003884.OF,003885.OF,003956.OF,003957.OF,003958.OF,003984.OF,003985.OF,003986.OF,004040.OF,004041.OF,004069.OF,004070.OF,004075.OF,004098.OF,004189.OF,004190.OF,004191.OF,004192.OF,004193.OF,004194.OF,004195.OF,004241.OF,004292.OF,004342.OF,004343.OF,004344.OF,004345.OF,004346.OF,004347.OF,004348.OF,004350.OF,004352.OF,004403.OF,004404.OF,004407.OF,004408.OF,004409.OF,004410.OF,004432.OF,004433.OF,004450.OF,004476.OF,004488.OF,004512.OF,004513.OF,004532.OF,004533.OF,004597.OF,004598.OF,004606.OF,004616.OF,004640.OF,004642.OF,004643.OF,004683.OF,004686.OF,004698.OF,004716.OF,004730.OF,004742.OF,004743.OF,004744.OF,004746.OF,004752.OF,004753.OF,004784.OF,004788.OF,004789.OF,004790.OF,004791.OF,004805.OF,004812.OF,004813.OF,004851.OF,004854.OF,004855.OF,004856.OF,004857.OF,004858.OF,004870.OF,004874.OF,004875.OF,004876.OF,004891.OF,004892.OF,004925.OF,004945.OF,004995.OF,004996.OF,004997.OF,005009.OF,005035.OF,005036.OF,005037.OF,005038.OF,005051.OF,005052.OF,005062.OF,005063.OF,005064.OF,005102.OF,005103.OF,005106.OF,005112.OF,005113.OF,005114.OF,005125.OF,005137.OF,005152.OF,005161.OF,005188.OF,005189.OF,005209.OF,005210.OF,005223.OF,005224.OF,005229.OF,005235.OF,005236.OF,005237.OF,005238.OF,005248.OF,005259.OF,005267.OF,005268.OF,005287.OF,005288.OF,005303.OF,005304.OF,005310.OF,005313.OF,005314.OF,005328.OF,005390.OF,005391.OF,005402.OF,005433.OF,005457.OF,005495.OF,005496.OF,005502.OF,005505.OF,005506.OF,005530.OF,005554.OF,005555.OF,005561.OF,005562.OF,005569.OF,005570.OF,005607.OF,005608.OF,005612.OF,005620.OF,005621.OF,005626.OF,005627.OF,005628.OF,005629.OF,005633.OF,005635.OF,005636.OF,005639.OF,005640.OF,005658.OF,005659.OF,005660.OF,005661.OF,005662.OF,005663.OF,005669.OF,005691.OF,005693.OF,005702.OF,005707.OF,005727.OF,005733.OF,005734.OF,005735.OF,005737.OF,005761.OF,005762.OF,005763.OF,005770.OF,005777.OF,005788.OF,005789.OF,005795.OF,005802.OF,005813.OF,005814.OF,005825.OF,005826.OF,005829.OF,005830.OF,005832.OF,005867.OF,005868.OF,005869.OF,005870.OF,005873.OF,005874.OF,005880.OF,005881.OF,005885.OF,005894.OF,005918.OF,005919.OF,005927.OF,005928.OF,005960.OF,005961.OF,005962.OF,005963.OF,005965.OF,005966.OF,005968.OF,005969.OF,005970.OF,005983.OF,005994.OF,005998.OF,005999.OF,006002.OF,006003.OF,006020.OF,006021.OF,006034.OF,006048.OF,006063.OF,006080.OF,006081.OF,006087.OF,006098.OF,006106.OF,006119.OF,006129.OF,006130.OF,006131.OF,006138.OF,006143.OF,006144.OF,006165.OF,006166.OF,006193.OF,006194.OF,006195.OF,006196.OF,006197.OF,006214.OF,006215.OF,006220.OF,006221.OF,006228.OF,006229.OF,006248.OF,006249.OF,006251.OF,006257.OF,006265.OF,006286.OF,006293.OF,006339.OF,006341.OF,006342.OF,006346.OF,006347.OF,006355.OF,006363.OF,006382.OF,006395.OF,006438.OF,006439.OF,006440.OF,006441.OF,006478.OF,006486.OF,006487.OF,006502.OF,006503.OF,006524.OF,006525.OF,006560.OF,006561.OF,006569.OF,006593.OF,006594.OF,006600.OF,006603.OF,006604.OF,006605.OF,006611.OF,006614.OF,006656.OF,006657.OF,006658.OF,006659.OF,006671.OF,006675.OF,006676.OF,006682.OF,006687.OF,006688.OF,006692.OF,006693.OF,006697.OF,006704.OF,006705.OF,006712.OF,006713.OF,006724.OF,006729.OF,006730.OF,006733.OF,006743.OF,006744.OF,006748.OF,006751.OF,006756.OF,006757.OF,006781.OF,006783.OF,006784.OF,006786.OF,006787.OF,006803.OF,006809.OF,006810.OF,006818.OF,006819.OF,006868.OF,006909.OF,006910.OF,006911.OF,006912.OF,006923.OF,006924.OF,006926.OF,006928.OF,006937.OF,006938.OF,006939.OF,006957.OF,006981.OF,007005.OF,007028.OF,007029.OF,007039.OF,007044.OF,007045.OF,007076.OF,007077.OF,007089.OF,007096.OF,007107.OF,007108.OF,007130.OF,007133.OF,007134.OF,007135.OF,007136.OF,007137.OF,007138.OF,007143.OF,007144.OF,007153.OF,007154.OF,007178.OF,007191.OF,007216.OF,007223.OF,007230.OF,007275.OF,007276.OF,007287.OF,007288.OF,007300.OF,007301.OF,007339.OF,007354.OF,007357.OF,007379.OF,007380.OF,007386.OF,007388.OF,007389.OF,007397.OF,007404.OF,007405.OF,007410.OF,007411.OF,007413.OF,007431.OF,007448.OF,007464.OF,007465.OF,007466.OF,007467.OF,007470.OF,007471.OF,007472.OF,007473.OF,007474.OF,007475.OF,007505.OF,007506.OF,007512.OF,007523.OF,007524.OF,007531.OF,007538.OF,007539.OF,007541.OF,007548.OF,007571.OF,007580.OF,007586.OF,007590.OF,007593.OF,007594.OF,007605.OF,007606.OF,007657.OF,007658.OF,007664.OF,007665.OF,007671.OF,007672.OF,007674.OF,007675.OF,007685.OF,007737.OF,007750.OF,007751.OF,007760.OF,007782.OF,007783.OF,007784.OF,007785.OF,007786.OF,007787.OF,007788.OF,007789.OF,007792.OF,007793.OF,007794.OF,007795.OF,007796.OF,007797.OF,007799.OF,007800.OF,007801.OF,007804.OF,007806.OF,007807.OF,007809.OF,007810.OF,007815.OF,007816.OF,007817.OF,007818.OF,007831.OF,007832.OF,007839.OF,007840.OF,007842.OF,007843.OF,007853.OF,007856.OF,007857.OF,007873.OF,007874.OF,007882.OF,007883.OF,007885.OF,007886.OF,007903.OF,007932.OF,007943.OF,007950.OF,007952.OF,007965.OF,007966.OF,007983.OF,007984.OF,007992.OF,007993.OF,007994.OF,007995.OF,008001.OF,008009.OF,008020.OF,008021.OF,008052.OF,008053.OF,008056.OF,008057.OF,008072.OF,008075.OF,008076.OF,008084.OF,008085.OF,008086.OF,008087.OF,008088.OF,008089.OF,008107.OF,008112.OF,008113.OF,008114.OF,008115.OF,008124.OF,008128.OF,008129.OF,008134.OF,008154.OF,008155.OF,008163.OF,008164.OF,008166.OF,008167.OF,008177.OF,008184.OF,008189.OF,008190.OF,008199.OF,008200.OF,008201.OF,008234.OF,008236.OF,008237.OF,008238.OF,008239.OF,008240.OF,008241.OF,008249.OF,008250.OF,008258.OF,008259.OF,008261.OF,008262.OF,008264.OF,008265.OF,008279.OF,008280.OF,008281.OF,008282.OF,008283.OF,008286.OF,008291.OF,008292.OF,008294.OF,008295.OF,008298.OF,008299.OF,008303.OF,008304.OF,008314.OF,008315.OF,008326.OF,008327.OF,008382.OF,008390.OF,008391.OF,008396.OF,008397.OF,008399.OF,008400.OF,008407.OF,008408.OF,008412.OF,008413.OF,008445.OF,008519.OF,008520.OF,008537.OF,008538.OF,008551.OF,008552.OF,008585.OF,008586.OF,008590.OF,008591.OF,008592.OF,008593.OF,008618.OF,008619.OF,008682.OF,008709.OF,008713.OF,008714.OF,008736.OF,008737.OF,008749.OF,008750.OF,008768.OF,008769.OF,008776.OF,008777.OF,008778.OF,008779.OF,008832.OF,008854.OF,008855.OF,008860.OF,008871.OF,008872.OF,008887.OF,008888.OF,008889.OF,008898.OF,008899.OF,008907.OF,008908.OF,008916.OF,008917.OF,008923.OF,008924.OF,008926.OF,008927.OF,008928.OF,008929.OF,008934.OF,008935.OF,008944.OF,008945.OF,008958.OF,008973.OF,008975.OF,008976.OF,009004.OF,009012.OF,009013.OF,009017.OF,009043.OF,009046.OF,009047.OF,009051.OF,009052.OF,009059.OF,009060.OF,009067.OF,009068.OF,009069.OF,009070.OF,009079.OF,009080.OF,009106.OF,009107.OF,009116.OF,009117.OF,009126.OF,009127.OF,009147.OF,009162.OF,009163.OF,009179.OF,009180.OF,009194.OF,009195.OF,009208.OF,009239.OF,009240.OF,009265.OF,009300.OF,009329.OF,009336.OF,009337.OF,009341.OF,009374.OF,009375.OF,009379.OF,009384.OF,009398.OF,009414.OF,009439.OF,009472.OF,009476.OF,009479.OF,009480.OF,009491.OF,009492.OF,009502.OF,009503.OF,009513.OF,009514.OF,009537.OF,009538.OF,009550.OF,009551.OF,009557.OF,009601.OF,009602.OF,009608.OF,009609.OF,009613.OF,009614.OF,009658.OF,009726.OF,009727.OF,009733.OF,009734.OF,009775.OF,009794.OF,009805.OF,009852.OF,009860.OF,009863.OF,009864.OF,009865.OF,009874.OF,009875.OF,009876.OF,009877.OF,009881.OF,009891.OF,009896.OF,009897.OF,009898.OF,009899.OF,009904.OF,009905.OF,009954.OF,009981.OF,009982.OF,009988.OF,010001.OF,010002.OF,010003.OF,010004.OF,010010.OF,010013.OF,010024.OF,010066.OF,010067.OF,010112.OF,010113.OF,010120.OF,010121.OF,010128.OF,010129.OF,010141.OF,010142.OF,010144.OF,010153.OF,010154.OF,010157.OF,010158.OF,010160.OF,010161.OF,010162.OF,010183.OF,010196.OF,010197.OF,010202.OF,010203.OF,010204.OF,010210.OF,010224.OF,010235.OF,010236.OF,010245.OF,010253.OF,010321.OF,010322.OF,010331.OF,010332.OF,010341.OF,010342.OF,010349.OF,010351.OF,010352.OF,010355.OF,010356.OF,010361.OF,010362.OF,010364.OF,010365.OF,010366.OF,010387.OF,010388.OF,010391.OF,010392.OF,010409.OF,010419.OF,010420.OF,010432.OF,010434.OF,010469.OF,010470.OF,010495.OF,010496.OF,010505.OF,010524.OF,010531.OF,010556.OF,010561.OF,010572.OF,010573.OF,010585.OF,010586.OF,010592.OF,010593.OF,010673.OF,010674.OF,010677.OF,010680.OF,010681.OF,010685.OF,010687.OF,010703.OF,010704.OF,010709.OF,010710.OF,010736.OF,010737.OF,010749.OF,010750.OF,010769.OF,010770.OF,010771.OF,010772.OF,010777.OF,010778.OF,010785.OF,010786.OF,010801.OF,010802.OF,010805.OF,010806.OF,010812.OF,010815.OF,010854.OF,010855.OF,010872.OF,010873.OF,010892.OF,010893.OF,010896.OF,010897.OF,010908.OF,010909.OF,010953.OF,010954.OF,010955.OF,010956.OF,010957.OF,010961.OF,010989.OF,010990.OF,010992.OF,010993.OF,010998.OF,011013.OF,011014.OF,011035.OF,011036.OF,011040.OF,011041.OF,011042.OF,011043.OF,011066.OF,011077.OF,011102.OF,011103.OF,011104.OF,011107.OF,011108.OF,011122.OF,011123.OF,011125.OF,011126.OF,011132.OF,011133.OF,011142.OF,011143.OF,011146.OF,011147.OF,011149.OF,011150.OF,011179.OF,011180.OF,011205.OF,011220.OF,011221.OF,011224.OF,011225.OF,011229.OF,011230.OF,011236.OF,011237.OF,011254.OF,011255.OF,011269.OF,011270.OF,011285.OF,011286.OF,011295.OF,011316.OF,011317.OF,011319.OF,011320.OF,011321.OF,011322.OF,011323.OF,011326.OF,011328.OF,011329.OF,011349.OF,011350.OF,011373.OF,011374.OF,011377.OF,011378.OF,011384.OF,011385.OF,011418.OF,011419.OF,011424.OF,011425.OF,011426.OF,011457.OF,011463.OF,011470.OF,011473.OF,011476.OF,011478.OF,011498.OF,011499.OF,011506.OF,011507.OF,011512.OF,011513.OF,011545.OF,011546.OF,011589.OF,011590.OF,011592.OF,011601.OF,011602.OF,011607.OF,011608.OF,011609.OF,011610.OF,011611.OF,011612.OF,011613.OF,011614.OF,011615.OF,011630.OF,011631.OF,011645.OF,011646.OF,011651.OF,011652.OF,011667.OF,011668.OF,011685.OF,011686.OF,011722.OF,011723.OF,011726.OF,011793.OF,011824.OF,011825.OF,011832.OF,011833.OF,011836.OF,011839.OF,011840.OF,011853.OF,011854.OF,011860.OF,011861.OF,011882.OF,011883.OF,011923.OF,011926.OF,011934.OF,011935.OF,011966.OF,011967.OF,011971.OF,011972.OF,012040.OF,012041.OF,012042.OF,012043.OF,012044.OF,012045.OF,012046.OF,012080.OF,012081.OF,012116.OF,012117.OF,012126.OF,012127.OF,012151.OF,012152.OF,012157.OF,012158.OF,012179.OF,012180.OF,012181.OF,012206.OF,012207.OF,012210.OF,012211.OF,012212.OF,012213.OF,012214.OF,012215.OF,012238.OF,012244.OF,012245.OF,012275.OF,012276.OF,012315.OF,012316.OF,012319.OF,012320.OF,012321.OF,012322.OF,012323.OF,012326.OF,012327.OF,012340.OF,012341.OF,012362.OF,012363.OF,012364.OF,012365.OF,012371.OF,012372.OF,012394.OF,012401.OF,012402.OF,012405.OF,012414.OF,012415.OF,012416.OF,012417.OF,012419.OF,012445.OF,012461.OF,012462.OF,012496.OF,012497.OF,012498.OF,012499.OF,012503.OF,012504.OF,012516.OF,012517.OF,012537.OF,012538.OF,012543.OF,012544.OF,012547.OF,012548.OF,012549.OF,012550.OF,012551.OF,012552.OF,012553.OF,012554.OF,012555.OF,012559.OF,012560.OF,012561.OF,012562.OF,012590.OF,012596.OF,012597.OF,012598.OF,012599.OF,012600.OF,012605.OF,012606.OF,012619.OF,012620.OF,012629.OF,012630.OF,012634.OF,012635.OF,012636.OF,012637.OF,012643.OF,012644.OF,012645.OF,012646.OF,012655.OF,012658.OF,012663.OF,012664.OF,012679.OF,012680.OF,012694.OF,012695.OF,012696.OF,012697.OF,012698.OF,012699.OF,012700.OF,012708.OF,012709.OF,012712.OF,012713.OF,012717.OF,012718.OF,012722.OF,012723.OF,012724.OF,012725.OF,012728.OF,012729.OF,012730.OF,012731.OF,012733.OF,012734.OF,012737.OF,012738.OF,012754.OF,012755.OF,012756.OF,012757.OF,012759.OF,012760.OF,012761.OF,012762.OF,012763.OF,012764.OF,012768.OF,012769.OF,012781.OF,012782.OF,012800.OF,012801.OF,012802.OF,012808.OF,012809.OF,012810.OF,012811.OF,012831.OF,012832.OF,012835.OF,012836.OF,012837.OF,012838.OF,012842.OF,012854.OF,012855.OF,012857.OF,012862.OF,012863.OF,012872.OF,012873.OF,012874.OF,012875.OF,012882.OF,012883.OF,012884.OF,012885.OF,012886.OF,012894.OF,012895.OF,012898.OF,012899.OF,012900.OF,012901.OF,012907.OF,012908.OF,012911.OF,012912.OF,012926.OF,012927.OF,012930.OF,012961.OF,012962.OF,012969.OF,012970.OF,012973.OF,012974.OF,012989.OF,012990.OF,013004.OF,013005.OF,013011.OF,013012.OF,013013.OF,013014.OF,013019.OF,013020.OF,013031.OF,013032.OF,013035.OF,013048.OF,013049.OF,013050.OF,013053.OF,013054.OF,013074.OF,013080.OF,013081.OF,013082.OF,013083.OF,013084.OF,013093.OF,013094.OF,013105.OF,013106.OF,013119.OF,013120.OF,013121.OF,013122.OF,013125.OF,013126.OF,013129.OF,013130.OF,013132.OF,013133.OF,013134.OF,013151.OF,013152.OF,013162.OF,013163.OF,013174.OF,013177.OF,013178.OF,013179.OF,013180.OF,013188.OF,013195.OF,013196.OF,013218.OF,013219.OF,013233.OF,013234.OF,013242.OF,013262.OF,013273.OF,013275.OF,013276.OF,013277.OF,013278.OF,013286.OF,013291.OF,013292.OF,013298.OF,013299.OF,013302.OF,013303.OF,013304.OF,013305.OF,013306.OF,013307.OF,013310.OF,013311.OF,013313.OF,013314.OF,013315.OF,013316.OF,013317.OF,013318.OF,013319.OF,013320.OF,013330.OF,013331.OF,013332.OF,013339.OF,013340.OF,013355.OF,013413.OF,013414.OF,013415.OF,013416.OF,013437.OF,013438.OF,013442.OF,013443.OF,013444.OF,013445.OF,013446.OF,013447.OF,013448.OF,013453.OF,013465.OF,013466.OF,013471.OF,013472.OF,013473.OF,013474.OF,013475.OF,013476.OF,013477.OF,013478.OF,013479.OF,013480.OF,013481.OF,013490.OF,013491.OF,013502.OF,013503.OF,013505.OF,013511.OF,013596.OF,013597.OF,013601.OF,013602.OF,013605.OF,013606.OF,013613.OF,013622.OF,013639.OF,013640.OF,013641.OF,013642.OF,013802.OF,013803.OF,013810.OF,013811.OF,013816.OF,013817.OF,013833.OF,013834.OF,013869.OF,013870.OF,013873.OF,013874.OF,013878.OF,013879.OF,013881.OF,013899.OF,013900.OF,013907.OF,013908.OF,013918.OF,013919.OF,013922.OF,013923.OF,013926.OF,013927.OF,013940.OF,013941.OF,013942.OF,013943.OF,014018.OF,014019.OF,014028.OF,014046.OF,014052.OF,014062.OF,014063.OF,014064.OF,014099.OF,014100.OF,014110.OF,014111.OF,014117.OF,014118.OF,014125.OF,014126.OF,014128.OF,014129.OF,014130.OF,014133.OF,014134.OF,014153.OF,014154.OF,014155.OF,014156.OF,014163.OF,014172.OF,014173.OF,014174.OF,014185.OF,014186.OF,014187.OF,014188.OF,014191.OF,014192.OF,014193.OF,014194.OF,014198.OF,014201.OF,014202.OF,014218.OF,014219.OF,014226.OF,014227.OF,014237.OF,014238.OF,014256.OF,014299.OF,014300.OF,014303.OF,014304.OF,014331.OF,014332.OF,014344.OF,014345.OF,014364.OF,014372.OF,014376.OF,014377.OF,014380.OF,014382.OF,014406.OF,014407.OF,014416.OF,014417.OF,014418.OF,014419.OF,014431.OF,014432.OF,014453.OF,014455.OF,014519.OF,014520.OF,014528.OF,014529.OF,014530.OF,014531.OF,014532.OF,014533.OF,014534.OF,014535.OF,014536.OF,014543.OF,014544.OF,014602.OF,014603.OF,014604.OF,014605.OF,014629.OF,014630.OF,014631.OF,014673.OF,014674.OF,014689.OF,014690.OF,014736.OF,014737.OF,014758.OF,014776.OF,014777.OF,014820.OF,014821.OF,014831.OF,014832.OF,014863.OF,014864.OF,014866.OF,014902.OF,014903.OF,014906.OF,014907.OF,014908.OF,014909.OF,014930.OF,014932.OF,014937.OF,014942.OF,014943.OF,014967.OF,014979.OF,014980.OF,014983.OF,014984.OF,014985.OF,014994.OF,014997.OF,015001.OF,015002.OF,015037.OF,015038.OF,015040.OF,015042.OF,015046.OF,015048.OF,015052.OF,015053.OF,015060.OF,015075.OF,015085.OF,015086.OF,015114.OF,015115.OF,015123.OF,015124.OF,015139.OF,015140.OF,015157.OF,015159.OF,015170.OF,015171.OF,015172.OF,015176.OF,015177.OF,015178.OF,015194.OF,015195.OF,015196.OF,015197.OF,015198.OF,015199.OF,015225.OF,015226.OF,015278.OF,015279.OF,015328.OF,015329.OF,015336.OF,015337.OF,015364.OF,015387.OF,015388.OF,015395.OF,015468.OF,015475.OF,015484.OF,020011.OF,020021.OF,040002.OF,040180.OF,040190.OF,050002.OF,050013.OF,050021.OF,050024.OF,070023.OF,070030.OF,070039.OF,090010.OF,090012.OF,100032.OF,100038.OF,100053.OF,110003.OF,110019.OF,110020.OF,110021.OF,110022.OF,110026.OF,110030.OF,159601.OF,159602.OF,159606.OF,159608.OF,159610.OF,159611.OF,159613.OF,159618.OF,159619.OF,159701.OF,159702.OF,159703.OF,159706.OF,159707.OF,159708.OF,159709.OF,159710.OF,159711.OF,159712.OF,159713.OF,159715.OF,159716.OF,159717.OF,159718.OF,159719.OF,159720.OF,159721.OF,159723.OF,159725.OF,159726.OF,159728.OF,159729.OF,159730.OF,159731.OF,159732.OF,159733.OF,159735.OF,159736.OF,159738.OF,159739.OF,159743.OF,159745.OF,159748.OF,159751.OF,159752.OF,159755.OF,159757.OF,159758.OF,159760.OF,159761.OF,159763.OF,159766.OF,159767.OF,159768.OF,159769.OF,159770.OF,159773.OF,159775.OF,159776.OF,159777.OF,159778.OF,159779.OF,159780.OF,159781.OF,159782.OF,159783.OF,159786.OF,159787.OF,159788.OF,159789.OF,159790.OF,159791.OF,159792.OF,159793.OF,159795.OF,159796.OF,159801.OF,159804.OF,159805.OF,159806.OF,159807.OF,159808.OF,159810.OF,159811.OF,159813.OF,159814.OF,159815.OF,159819.OF,159820.OF,159821.OF,159824.OF,159825.OF,159827.OF,159828.OF,159835.OF,159836.OF,159837.OF,159838.OF,159839.OF,159840.OF,159841.OF,159842.OF,159843.OF,159845.OF,159847.OF,159848.OF,159849.OF,159851.OF,159852.OF,159853.OF,159855.OF,159856.OF,159857.OF,159858.OF,159859.OF,159861.OF,159862.OF,159863.OF,159864.OF,159865.OF,159867.OF,159869.OF,159870.OF,159871.OF,159872.OF,159873.OF,159875.OF,159876.OF,159877.OF,159880.OF,159881.OF,159883.OF,159885.OF,159886.OF,159887.OF,159888.OF,159889.OF,159890.OF,159891.OF,159895.OF,159896.OF,159897.OF,159898.OF,159899.OF,159901.OF,159902.OF,159903.OF,159905.OF,159906.OF,159907.OF,159908.OF,159909.OF,159910.OF,159912.OF,159913.OF,159915.OF,159916.OF,159918.OF,159919.OF,159922.OF,159923.OF,159925.OF,159928.OF,159929.OF,159930.OF,159931.OF,159932.OF,159933.OF,159935.OF,159936.OF,159938.OF,159939.OF,159940.OF,159943.OF,159944.OF,159945.OF,159948.OF,159949.OF,159951.OF,159952.OF,159955.OF,159956.OF,159957.OF,159958.OF,159959.OF,159960.OF,159961.OF,159962.OF,159964.OF,159965.OF,159966.OF,159967.OF,159968.OF,159969.OF,159970.OF,159971.OF,159973.OF,159974.OF,159975.OF,159976.OF,159977.OF,159978.OF,159979.OF,159982.OF,159983.OF,159984.OF,159986.OF,159987.OF,159990.OF,159991.OF,159992.OF,159993.OF,159994.OF,159995.OF,159996.OF,159997.OF,159998.OF,159999.OF,160119.OF,160127.OF,160133.OF,160135.OF,160137.OF,160144.OF,160218.OF,160219.OF,160221.OF,160222.OF,160223.OF,160224.OF,160225.OF,160322.OF,160418.OF,160419.OF,160420.OF,160422.OF,160424.OF,160516.OF,160517.OF,160615.OF,160616.OF,160620.OF,160625.OF,160626.OF,160628.OF,160629.OF,160630.OF,160631.OF,160632.OF,160633.OF,160634.OF,160635.OF,160636.OF,160637.OF,160638.OF,160639.OF,160643.OF,160646.OF,160706.OF,160716.OF,160724.OF,160725.OF,160806.OF,160807.OF,160814.OF,160919.OF,160925.OF,161017.OF,161022.OF,161024.OF,161025.OF,161026.OF,161027.OF,161028.OF,161029.OF,161030.OF,161031.OF,161032.OF,161033.OF,161035.OF,161036.OF,161037.OF,161039.OF,161118.OF,161121.OF,161122.OF,161123.OF,161211.OF,161217.OF,161227.OF,161604.OF,161607.OF,161612.OF,161613.OF,161628.OF,161631.OF,161715.OF,161718.OF,161720.OF,161721.OF,161723.OF,161724.OF,161725.OF,161726.OF,161811.OF,161812.OF,161816.OF,161907.OF,162107.OF,162208.OF,162213.OF,162216.OF,162307.OF,162412.OF,162413.OF,162416.OF,162509.OF,162510.OF,162711.OF,162714.OF,162907.OF,163109.OF,163110.OF,163111.OF,163113.OF,163114.OF,163115.OF,163116.OF,163118.OF,163209.OF,163407.OF,163808.OF,163821.OF,164205.OF,164304.OF,164401.OF,164402.OF,164508.OF,164809.OF,164811.OF,164818.OF,164825.OF,164905.OF,164908.OF,165309.OF,165310.OF,165312.OF,165511.OF,165515.OF,165519.OF,165520.OF,165521.OF,165522.OF,165523.OF,165524.OF,165525.OF,165806.OF,165810.OF,166402.OF,166802.OF,167301.OF,167302.OF,167503.OF,167506.OF,167507.OF,167601.OF,167702.OF,167703.OF,168203.OF,168204.OF,168205.OF,168701.OF,168702.OF,180003.OF,200002.OF,202015.OF,202017.OF,202021.OF,202025.OF,202211.OF,206012.OF,213010.OF,217016.OF,217017.OF,217019.OF,217027.OF,233010.OF,240014.OF,240016.OF,240019.OF,257060.OF,270010.OF,270026.OF,290010.OF,310318.OF,310398.OF,320010.OF,320014.OF,320020.OF,320022.OF,340006.OF,360001.OF,370023.OF,376510.OF,399001.OF,399011.OF,410008.OF,410010.OF,450008.OF,450009.OF,460220.OF,460300.OF,470007.OF,470068.OF,481009.OF,481012.OF,501005.OF,501006.OF,501007.OF,501008.OF,501009.OF,501010.OF,501011.OF,501012.OF,501016.OF,501019.OF,501025.OF,501029.OF,501030.OF,501031.OF,501036.OF,501037.OF,501043.OF,501045.OF,501047.OF,501048.OF,501050.OF,501057.OF,501058.OF,501059.OF,501060.OF,501061.OF,501067.OF,501068.OF,501069.OF,501086.OF,501089.OF,501090.OF,501219.OF,501301.OF,501302.OF,501303.OF,501305.OF,501306.OF,501307.OF,501308.OF,501309.OF,501310.OF,501311.OF,502000.OF,502003.OF,502006.OF,502010.OF,502013.OF,502020.OF,502023.OF,502040.OF,502048.OF,502053.OF,502056.OF,510010.OF,510020.OF,510030.OF,510050.OF,510060.OF,510090.OF,510100.OF,510110.OF,510120.OF,510130.OF,510150.OF,510160.OF,510170.OF,510180.OF,510190.OF,510200.OF,510210.OF,510220.OF,510230.OF,510270.OF,510290.OF,510300.OF,510310.OF,510330.OF,510350.OF,510360.OF,510370.OF,510380.OF,510390.OF,510410.OF,510440.OF,510500.OF,510510.OF,510530.OF,510550.OF,510560.OF,510570.OF,510580.OF,510590.OF,510600.OF,510630.OF,510650.OF,510660.OF,510680.OF,510710.OF,510760.OF,510770.OF,510800.OF,510810.OF,510850.OF,510880.OF,510990.OF,512000.OF,512010.OF,512040.OF,512070.OF,512090.OF,512100.OF,512120.OF,512150.OF,512160.OF,512170.OF,512180.OF,512190.OF,512200.OF,512220.OF,512260.OF,512280.OF,512290.OF,512320.OF,512330.OF,512360.OF,512380.OF,512390.OF,512400.OF,512480.OF,512500.OF,512510.OF,512520.OF,512530.OF,512550.OF,512560.OF,512570.OF,512580.OF,512590.OF,512600.OF,512610.OF,512640.OF,512650.OF,512660.OF,512670.OF,512680.OF,512690.OF,512700.OF,512710.OF,512720.OF,512730.OF,512750.OF,512760.OF,512770.OF,512780.OF,512800.OF,512810.OF,512820.OF,512870.OF,512880.OF,512890.OF,512900.OF,512910.OF,512920.OF,512930.OF,512950.OF,512960.OF,512970.OF,512980.OF,512990.OF,513020.OF,513070.OF,513090.OF,513150.OF,513160.OF,513200.OF,513230.OF,513320.OF,513550.OF,513590.OF,513660.OF,513680.OF,513690.OF,513700.OF,513770.OF,513860.OF,513900.OF,513960.OF,513980.OF,513990.OF,515000.OF,515010.OF,515020.OF,515030.OF,515050.OF,515060.OF,515070.OF,515080.OF,515090.OF,515100.OF,515110.OF,515120.OF,515130.OF,515150.OF,515160.OF,515170.OF,515180.OF,515190.OF,515200.OF,515210.OF,515220.OF,515230.OF,515250.OF,515260.OF,515280.OF,515290.OF,515300.OF,515310.OF,515320.OF,515330.OF,515350.OF,515360.OF,515380.OF,515390.OF,515400.OF,515450.OF,515500.OF,515510.OF,515520.OF,515530.OF,515550.OF,515560.OF,515570.OF,515580.OF,515590.OF,515600.OF,515610.OF,515630.OF,515650.OF,515660.OF,515670.OF,515680.OF,515700.OF,515710.OF,515750.OF,515760.OF,515770.OF,515780.OF,515790.OF,515800.OF,515810.OF,515820.OF,515850.OF,515860.OF,515870.OF,515880.OF,515890.OF,515900.OF,515910.OF,515920.OF,515930.OF,515950.OF,515960.OF,515980.OF,515990.OF,516000.OF,516010.OF,516020.OF,516050.OF,516060.OF,516070.OF,516080.OF,516090.OF,516100.OF,516110.OF,516120.OF,516130.OF,516150.OF,516160.OF,516180.OF,516190.OF,516200.OF,516210.OF,516220.OF,516260.OF,516270.OF,516290.OF,516300.OF,516310.OF,516320.OF,516330.OF,516350.OF,516360.OF,516380.OF,516390.OF,516480.OF,516500.OF,516510.OF,516520.OF,516530.OF,516550.OF,516560.OF,516570.OF,516580.OF,516590.OF,516600.OF,516610.OF,516620.OF,516630.OF,516640.OF,516650.OF,516660.OF,516670.OF,516680.OF,516690.OF,516700.OF,516710.OF,516720.OF,516730.OF,516750.OF,516760.OF,516770.OF,516780.OF,516790.OF,516800.OF,516810.OF,516820.OF,516830.OF,516850.OF,516860.OF,516880.OF,516890.OF,516900.OF,516910.OF,516920.OF,516930.OF,516950.OF,516960.OF,516970.OF,516980.OF,517000.OF,517010.OF,517030.OF,517050.OF,517080.OF,517090.OF,517100.OF,517110.OF,517120.OF,517160.OF,517170.OF,517180.OF,517200.OF,517270.OF,517280.OF,517300.OF,517330.OF,517350.OF,517360.OF,517380.OF,517390.OF,517500.OF,517550.OF,517660.OF,517760.OF,517770.OF,517780.OF,517800.OF,517880.OF,517900.OF,517960.OF,519027.OF,519032.OF,519034.OF,519100.OF,519116.OF,519117.OF,519180.OF,519193.OF,519300.OF,519606.OF,519671.OF,519673.OF,519677.OF,519686.OF,519706.OF,519714.OF,519935.OF,519965.OF,519975.OF,530010.OF,530015.OF,530018.OF,540006.OF,540007.OF,540008.OF,540009.OF,540010.OF,540012.OF,560000.OF,560050.OF,560100.OF,560500.OF,560560.OF,560600.OF,560660.OF,560800.OF,560880.OF,561100.OF,561120.OF,561130.OF,561300.OF,561310.OF,561350.OF,561500.OF,561550.OF,561600.OF,561800.OF,561900.OF,561910.OF,561990.OF,562300.OF,562360.OF,562500.OF,562510.OF,562520.OF,562800.OF,562880.OF,562900.OF,562910.OF,562950.OF,563000.OF,580008.OF,585001.OF,588000.OF,588050.OF,588060.OF,588080.OF,588090.OF,588150.OF,588180.OF,588280.OF,588300.OF,588310.OF,588330.OF,588360.OF,588380.OF,588390.OF,588400.OF,590007.OF,660008.OF,660011.OF,671030.OF,673090.OF,673100.OF,673101.OF,690008.OF,700002.OF,740101.OF,900029.OF,900030.OF,920002.OF,920003.OF,920922.OF,920923.OF,960000.OF,960022.OF,970041.OF,970042.OF,H161604.OF,H202017.OF','ths_fund_thscode_fund','')
    if data_fund.errorcode != 0:
        print('error:{}'.format(data_fund.errmsg))
    else:
        all_fund=THS_BD('000008.OF,000042.OF,000051.OF,000059.OF,000082.OF,000176.OF,000248.OF,000309.OF,000311.OF,000312.OF,000313.OF,000326.OF,000368.OF,000373.OF,000376.OF,000409.OF,000411.OF,000418.OF,000457.OF,000471.OF,000478.OF,000512.OF,000513.OF,000524.OF,000549.OF,000577.OF,000586.OF,000592.OF,000594.OF,000596.OF,000613.OF,000628.OF,000656.OF,000688.OF,000696.OF,000697.OF,000711.OF,000729.OF,000746.OF,000751.OF,000756.OF,000761.OF,000778.OF,000780.OF,000793.OF,000803.OF,000826.OF,000827.OF,000828.OF,000831.OF,000835.OF,000854.OF,000866.OF,000867.OF,000884.OF,000893.OF,000913.OF,000916.OF,000925.OF,000942.OF,000948.OF,000950.OF,000955.OF,000960.OF,000961.OF,000962.OF,000968.OF,000971.OF,000974.OF,000975.OF,000978.OF,000979.OF,000985.OF,000991.OF,000996.OF,001008.OF,001009.OF,001015.OF,001016.OF,001027.OF,001028.OF,001036.OF,001039.OF,001040.OF,001042.OF,001043.OF,001044.OF,001047.OF,001048.OF,001050.OF,001051.OF,001052.OF,001054.OF,001064.OF,001070.OF,001072.OF,001097.OF,001104.OF,001105.OF,001113.OF,001126.OF,001133.OF,001149.OF,001158.OF,001162.OF,001163.OF,001166.OF,001167.OF,001171.OF,001178.OF,001180.OF,001186.OF,001188.OF,001193.OF,001195.OF,001208.OF,001214.OF,001223.OF,001230.OF,001236.OF,001237.OF,001241.OF,001242.OF,001243.OF,001245.OF,001277.OF,001291.OF,001313.OF,001319.OF,001344.OF,001351.OF,001361.OF,001396.OF,001397.OF,001404.OF,001409.OF,001410.OF,001416.OF,001420.OF,001421.OF,001426.OF,001455.OF,001469.OF,001473.OF,001476.OF,001482.OF,001490.OF,001496.OF,001520.OF,001521.OF,001528.OF,001539.OF,001541.OF,001542.OF,001548.OF,001549.OF,001550.OF,001551.OF,001552.OF,001553.OF,001556.OF,001557.OF,001576.OF,001577.OF,001579.OF,001583.OF,001588.OF,001589.OF,001592.OF,001593.OF,001594.OF,001595.OF,001605.OF,001616.OF,001617.OF,001618.OF,001626.OF,001628.OF,001629.OF,001630.OF,001631.OF,001632.OF,001637.OF,001638.OF,001643.OF,001644.OF,001645.OF,001651.OF,001672.OF,001677.OF,001678.OF,001685.OF,001692.OF,001703.OF,001705.OF,001707.OF,001714.OF,001717.OF,001718.OF,001719.OF,001725.OF,001726.OF,001733.OF,001736.OF,001749.OF,001764.OF,001766.OF,001781.OF,001790.OF,001825.OF,001849.OF,001877.OF,001878.OF,001879.OF,001899.OF,001915.OF,001917.OF,001938.OF,001956.OF,001974.OF,001975.OF,002063.OF,002076.OF,002121.OF,002168.OF,002199.OF,002210.OF,002229.OF,002236.OF,002289.OF,002300.OF,002310.OF,002311.OF,002315.OF,002316.OF,002332.OF,002333.OF,002334.OF,002335.OF,002385.OF,002387.OF,002510.OF,002556.OF,002595.OF,002621.OF,002656.OF,002670.OF,002671.OF,002697.OF,002861.OF,002871.OF,002872.OF,002900.OF,002903.OF,002906.OF,002907.OF,002952.OF,002974.OF,002977.OF,002978.OF,002979.OF,002980.OF,002982.OF,002984.OF,002987.OF,003015.OF,003016.OF,003017.OF,003053.OF,003054.OF,003069.OF,003070.OF,003145.OF,003184.OF,003190.OF,003191.OF,003194.OF,003230.OF,003231.OF,003232.OF,003233.OF,003298.OF,003299.OF,003318.OF,003359.OF,003366.OF,003416.OF,003475.OF,003548.OF,003578.OF,003579.OF,003624.OF,003625.OF,003634.OF,003646.OF,003647.OF,003745.OF,003760.OF,003761.OF,003765.OF,003766.OF,003834.OF,003853.OF,003865.OF,003876.OF,003884.OF,003885.OF,003956.OF,003957.OF,003958.OF,003984.OF,003985.OF,003986.OF,004040.OF,004041.OF,004069.OF,004070.OF,004075.OF,004098.OF,004189.OF,004190.OF,004191.OF,004192.OF,004193.OF,004194.OF,004195.OF,004241.OF,004292.OF,004342.OF,004343.OF,004344.OF,004345.OF,004346.OF,004347.OF,004348.OF,004350.OF,004352.OF,004403.OF,004404.OF,004407.OF,004408.OF,004409.OF,004410.OF,004432.OF,004433.OF,004450.OF,004476.OF,004488.OF,004512.OF,004513.OF,004532.OF,004533.OF,004597.OF,004598.OF,004606.OF,004616.OF,004640.OF,004642.OF,004643.OF,004683.OF,004686.OF,004698.OF,004716.OF,004730.OF,004742.OF,004743.OF,004744.OF,004746.OF,004752.OF,004753.OF,004784.OF,004788.OF,004789.OF,004790.OF,004791.OF,004805.OF,004812.OF,004813.OF,004851.OF,004854.OF,004855.OF,004856.OF,004857.OF,004858.OF,004870.OF,004874.OF,004875.OF,004876.OF,004891.OF,004892.OF,004925.OF,004945.OF,004995.OF,004996.OF,004997.OF,005009.OF,005035.OF,005036.OF,005037.OF,005038.OF,005051.OF,005052.OF,005062.OF,005063.OF,005064.OF,005102.OF,005103.OF,005106.OF,005112.OF,005113.OF,005114.OF,005125.OF,005137.OF,005152.OF,005161.OF,005188.OF,005189.OF,005209.OF,005210.OF,005223.OF,005224.OF,005229.OF,005235.OF,005236.OF,005237.OF,005238.OF,005248.OF,005259.OF,005267.OF,005268.OF,005287.OF,005288.OF,005303.OF,005304.OF,005310.OF,005313.OF,005314.OF,005328.OF,005390.OF,005391.OF,005402.OF,005433.OF,005457.OF,005495.OF,005496.OF,005502.OF,005505.OF,005506.OF,005530.OF,005554.OF,005555.OF,005561.OF,005562.OF,005569.OF,005570.OF,005607.OF,005608.OF,005612.OF,005620.OF,005621.OF,005626.OF,005627.OF,005628.OF,005629.OF,005633.OF,005635.OF,005636.OF,005639.OF,005640.OF,005658.OF,005659.OF,005660.OF,005661.OF,005662.OF,005663.OF,005669.OF,005691.OF,005693.OF,005702.OF,005707.OF,005727.OF,005733.OF,005734.OF,005735.OF,005737.OF,005761.OF,005762.OF,005763.OF,005770.OF,005777.OF,005788.OF,005789.OF,005795.OF,005802.OF,005813.OF,005814.OF,005825.OF,005826.OF,005829.OF,005830.OF,005832.OF,005867.OF,005868.OF,005869.OF,005870.OF,005873.OF,005874.OF,005880.OF,005881.OF,005885.OF,005894.OF,005918.OF,005919.OF,005927.OF,005928.OF,005960.OF,005961.OF,005962.OF,005963.OF,005965.OF,005966.OF,005968.OF,005969.OF,005970.OF,005983.OF,005994.OF,005998.OF,005999.OF,006002.OF,006003.OF,006020.OF,006021.OF,006034.OF,006048.OF,006063.OF,006080.OF,006081.OF,006087.OF,006098.OF,006106.OF,006119.OF,006129.OF,006130.OF,006131.OF,006138.OF,006143.OF,006144.OF,006165.OF,006166.OF,006193.OF,006194.OF,006195.OF,006196.OF,006197.OF,006214.OF,006215.OF,006220.OF,006221.OF,006228.OF,006229.OF,006248.OF,006249.OF,006251.OF,006257.OF,006265.OF,006286.OF,006293.OF,006339.OF,006341.OF,006342.OF,006346.OF,006347.OF,006355.OF,006363.OF,006382.OF,006395.OF,006438.OF,006439.OF,006440.OF,006441.OF,006478.OF,006486.OF,006487.OF,006502.OF,006503.OF,006524.OF,006525.OF,006560.OF,006561.OF,006569.OF,006593.OF,006594.OF,006600.OF,006603.OF,006604.OF,006605.OF,006611.OF,006614.OF,006656.OF,006657.OF,006658.OF,006659.OF,006671.OF,006675.OF,006676.OF,006682.OF,006687.OF,006688.OF,006692.OF,006693.OF,006697.OF,006704.OF,006705.OF,006712.OF,006713.OF,006724.OF,006729.OF,006730.OF,006733.OF,006743.OF,006744.OF,006748.OF,006751.OF,006756.OF,006757.OF,006781.OF,006783.OF,006784.OF,006786.OF,006787.OF,006803.OF,006809.OF,006810.OF,006818.OF,006819.OF,006868.OF,006909.OF,006910.OF,006911.OF,006912.OF,006923.OF,006924.OF,006926.OF,006928.OF,006937.OF,006938.OF,006939.OF,006957.OF,006981.OF,007005.OF,007028.OF,007029.OF,007039.OF,007044.OF,007045.OF,007076.OF,007077.OF,007089.OF,007096.OF,007107.OF,007108.OF,007130.OF,007133.OF,007134.OF,007135.OF,007136.OF,007137.OF,007138.OF,007143.OF,007144.OF,007153.OF,007154.OF,007178.OF,007191.OF,007216.OF,007223.OF,007230.OF,007275.OF,007276.OF,007287.OF,007288.OF,007300.OF,007301.OF,007339.OF,007354.OF,007357.OF,007379.OF,007380.OF,007386.OF,007388.OF,007389.OF,007397.OF,007404.OF,007405.OF,007410.OF,007411.OF,007413.OF,007431.OF,007448.OF,007464.OF,007465.OF,007466.OF,007467.OF,007470.OF,007471.OF,007472.OF,007473.OF,007474.OF,007475.OF,007505.OF,007506.OF,007512.OF,007523.OF,007524.OF,007531.OF,007538.OF,007539.OF,007541.OF,007548.OF,007571.OF,007580.OF,007586.OF,007590.OF,007593.OF,007594.OF,007605.OF,007606.OF,007657.OF,007658.OF,007664.OF,007665.OF,007671.OF,007672.OF,007674.OF,007675.OF,007685.OF,007737.OF,007750.OF,007751.OF,007760.OF,007782.OF,007783.OF,007784.OF,007785.OF,007786.OF,007787.OF,007788.OF,007789.OF,007792.OF,007793.OF,007794.OF,007795.OF,007796.OF,007797.OF,007799.OF,007800.OF,007801.OF,007804.OF,007806.OF,007807.OF,007809.OF,007810.OF,007815.OF,007816.OF,007817.OF,007818.OF,007831.OF,007832.OF,007839.OF,007840.OF,007842.OF,007843.OF,007853.OF,007856.OF,007857.OF,007873.OF,007874.OF,007882.OF,007883.OF,007885.OF,007886.OF,007903.OF,007932.OF,007943.OF,007950.OF,007952.OF,007965.OF,007966.OF,007983.OF,007984.OF,007992.OF,007993.OF,007994.OF,007995.OF,008001.OF,008009.OF,008020.OF,008021.OF,008052.OF,008053.OF,008056.OF,008057.OF,008072.OF,008075.OF,008076.OF,008084.OF,008085.OF,008086.OF,008087.OF,008088.OF,008089.OF,008107.OF,008112.OF,008113.OF,008114.OF,008115.OF,008124.OF,008128.OF,008129.OF,008134.OF,008154.OF,008155.OF,008163.OF,008164.OF,008166.OF,008167.OF,008177.OF,008184.OF,008189.OF,008190.OF,008199.OF,008200.OF,008201.OF,008234.OF,008236.OF,008237.OF,008238.OF,008239.OF,008240.OF,008241.OF,008249.OF,008250.OF,008258.OF,008259.OF,008261.OF,008262.OF,008264.OF,008265.OF,008279.OF,008280.OF,008281.OF,008282.OF,008283.OF,008286.OF,008291.OF,008292.OF,008294.OF,008295.OF,008298.OF,008299.OF,008303.OF,008304.OF,008314.OF,008315.OF,008326.OF,008327.OF,008382.OF,008390.OF,008391.OF,008396.OF,008397.OF,008399.OF,008400.OF,008407.OF,008408.OF,008412.OF,008413.OF,008445.OF,008519.OF,008520.OF,008537.OF,008538.OF,008551.OF,008552.OF,008585.OF,008586.OF,008590.OF,008591.OF,008592.OF,008593.OF,008618.OF,008619.OF,008682.OF,008709.OF,008713.OF,008714.OF,008736.OF,008737.OF,008749.OF,008750.OF,008768.OF,008769.OF,008776.OF,008777.OF,008778.OF,008779.OF,008832.OF,008854.OF,008855.OF,008860.OF,008871.OF,008872.OF,008887.OF,008888.OF,008889.OF,008898.OF,008899.OF,008907.OF,008908.OF,008916.OF,008917.OF,008923.OF,008924.OF,008926.OF,008927.OF,008928.OF,008929.OF,008934.OF,008935.OF,008944.OF,008945.OF,008958.OF,008973.OF,008975.OF,008976.OF,009004.OF,009012.OF,009013.OF,009017.OF,009043.OF,009046.OF,009047.OF,009051.OF,009052.OF,009059.OF,009060.OF,009067.OF,009068.OF,009069.OF,009070.OF,009079.OF,009080.OF,009106.OF,009107.OF,009116.OF,009117.OF,009126.OF,009127.OF,009147.OF,009162.OF,009163.OF,009179.OF,009180.OF,009194.OF,009195.OF,009208.OF,009239.OF,009240.OF,009265.OF,009300.OF,009329.OF,009336.OF,009337.OF,009341.OF,009374.OF,009375.OF,009379.OF,009384.OF,009398.OF,009414.OF,009439.OF,009472.OF,009476.OF,009479.OF,009480.OF,009491.OF,009492.OF,009502.OF,009503.OF,009513.OF,009514.OF,009537.OF,009538.OF,009550.OF,009551.OF,009557.OF,009601.OF,009602.OF,009608.OF,009609.OF,009613.OF,009614.OF,009658.OF,009726.OF,009727.OF,009733.OF,009734.OF,009775.OF,009794.OF,009805.OF,009852.OF,009860.OF,009863.OF,009864.OF,009865.OF,009874.OF,009875.OF,009876.OF,009877.OF,009881.OF,009891.OF,009896.OF,009897.OF,009898.OF,009899.OF,009904.OF,009905.OF,009954.OF,009981.OF,009982.OF,009988.OF,010001.OF,010002.OF,010003.OF,010004.OF,010010.OF,010013.OF,010024.OF,010066.OF,010067.OF,010112.OF,010113.OF,010120.OF,010121.OF,010128.OF,010129.OF,010141.OF,010142.OF,010144.OF,010153.OF,010154.OF,010157.OF,010158.OF,010160.OF,010161.OF,010162.OF,010183.OF,010196.OF,010197.OF,010202.OF,010203.OF,010204.OF,010210.OF,010224.OF,010235.OF,010236.OF,010245.OF,010253.OF,010321.OF,010322.OF,010331.OF,010332.OF,010341.OF,010342.OF,010349.OF,010351.OF,010352.OF,010355.OF,010356.OF,010361.OF,010362.OF,010364.OF,010365.OF,010366.OF,010387.OF,010388.OF,010391.OF,010392.OF,010409.OF,010419.OF,010420.OF,010432.OF,010434.OF,010469.OF,010470.OF,010495.OF,010496.OF,010505.OF,010524.OF,010531.OF,010556.OF,010561.OF,010572.OF,010573.OF,010585.OF,010586.OF,010592.OF,010593.OF,010673.OF,010674.OF,010677.OF,010680.OF,010681.OF,010685.OF,010687.OF,010703.OF,010704.OF,010709.OF,010710.OF,010736.OF,010737.OF,010749.OF,010750.OF,010769.OF,010770.OF,010771.OF,010772.OF,010777.OF,010778.OF,010785.OF,010786.OF,010801.OF,010802.OF,010805.OF,010806.OF,010812.OF,010815.OF,010854.OF,010855.OF,010872.OF,010873.OF,010892.OF,010893.OF,010896.OF,010897.OF,010908.OF,010909.OF,010953.OF,010954.OF,010955.OF,010956.OF,010957.OF,010961.OF,010989.OF,010990.OF,010992.OF,010993.OF,010998.OF,011013.OF,011014.OF,011035.OF,011036.OF,011040.OF,011041.OF,011042.OF,011043.OF,011066.OF,011077.OF,011102.OF,011103.OF,011104.OF,011107.OF,011108.OF,011122.OF,011123.OF,011125.OF,011126.OF,011132.OF,011133.OF,011142.OF,011143.OF,011146.OF,011147.OF,011149.OF,011150.OF,011179.OF,011180.OF,011205.OF,011220.OF,011221.OF,011224.OF,011225.OF,011229.OF,011230.OF,011236.OF,011237.OF,011254.OF,011255.OF,011269.OF,011270.OF,011285.OF,011286.OF,011295.OF,011316.OF,011317.OF,011319.OF,011320.OF,011321.OF,011322.OF,011323.OF,011326.OF,011328.OF,011329.OF,011349.OF,011350.OF,011373.OF,011374.OF,011377.OF,011378.OF,011384.OF,011385.OF,011418.OF,011419.OF,011424.OF,011425.OF,011426.OF,011457.OF,011463.OF,011470.OF,011473.OF,011476.OF,011478.OF,011498.OF,011499.OF,011506.OF,011507.OF,011512.OF,011513.OF,011545.OF,011546.OF,011589.OF,011590.OF,011592.OF,011601.OF,011602.OF,011607.OF,011608.OF,011609.OF,011610.OF,011611.OF,011612.OF,011613.OF,011614.OF,011615.OF,011630.OF,011631.OF,011645.OF,011646.OF,011651.OF,011652.OF,011667.OF,011668.OF,011685.OF,011686.OF,011722.OF,011723.OF,011726.OF,011793.OF,011824.OF,011825.OF,011832.OF,011833.OF,011836.OF,011839.OF,011840.OF,011853.OF,011854.OF,011860.OF,011861.OF,011882.OF,011883.OF,011923.OF,011926.OF,011934.OF,011935.OF,011966.OF,011967.OF,011971.OF,011972.OF,012040.OF,012041.OF,012042.OF,012043.OF,012044.OF,012045.OF,012046.OF,012080.OF,012081.OF,012116.OF,012117.OF,012126.OF,012127.OF,012151.OF,012152.OF,012157.OF,012158.OF,012179.OF,012180.OF,012181.OF,012206.OF,012207.OF,012210.OF,012211.OF,012212.OF,012213.OF,012214.OF,012215.OF,012238.OF,012244.OF,012245.OF,012275.OF,012276.OF,012315.OF,012316.OF,012319.OF,012320.OF,012321.OF,012322.OF,012323.OF,012326.OF,012327.OF,012340.OF,012341.OF,012362.OF,012363.OF,012364.OF,012365.OF,012371.OF,012372.OF,012394.OF,012401.OF,012402.OF,012405.OF,012414.OF,012415.OF,012416.OF,012417.OF,012419.OF,012445.OF,012461.OF,012462.OF,012496.OF,012497.OF,012498.OF,012499.OF,012503.OF,012504.OF,012516.OF,012517.OF,012537.OF,012538.OF,012543.OF,012544.OF,012547.OF,012548.OF,012549.OF,012550.OF,012551.OF,012552.OF,012553.OF,012554.OF,012555.OF,012559.OF,012560.OF,012561.OF,012562.OF,012590.OF,012596.OF,012597.OF,012598.OF,012599.OF,012600.OF,012605.OF,012606.OF,012619.OF,012620.OF,012629.OF,012630.OF,012634.OF,012635.OF,012636.OF,012637.OF,012643.OF,012644.OF,012645.OF,012646.OF,012655.OF,012658.OF,012663.OF,012664.OF,012679.OF,012680.OF,012694.OF,012695.OF,012696.OF,012697.OF,012698.OF,012699.OF,012700.OF,012708.OF,012709.OF,012712.OF,012713.OF,012717.OF,012718.OF,012722.OF,012723.OF,012724.OF,012725.OF,012728.OF,012729.OF,012730.OF,012731.OF,012733.OF,012734.OF,012737.OF,012738.OF,012754.OF,012755.OF,012756.OF,012757.OF,012759.OF,012760.OF,012761.OF,012762.OF,012763.OF,012764.OF,012768.OF,012769.OF,012781.OF,012782.OF,012800.OF,012801.OF,012802.OF,012808.OF,012809.OF,012810.OF,012811.OF,012831.OF,012832.OF,012835.OF,012836.OF,012837.OF,012838.OF,012842.OF,012854.OF,012855.OF,012857.OF,012862.OF,012863.OF,012872.OF,012873.OF,012874.OF,012875.OF,012882.OF,012883.OF,012884.OF,012885.OF,012886.OF,012894.OF,012895.OF,012898.OF,012899.OF,012900.OF,012901.OF,012907.OF,012908.OF,012911.OF,012912.OF,012926.OF,012927.OF,012930.OF,012961.OF,012962.OF,012969.OF,012970.OF,012973.OF,012974.OF,012989.OF,012990.OF,013004.OF,013005.OF,013011.OF,013012.OF,013013.OF,013014.OF,013019.OF,013020.OF,013031.OF,013032.OF,013035.OF,013048.OF,013049.OF,013050.OF,013053.OF,013054.OF,013074.OF,013080.OF,013081.OF,013082.OF,013083.OF,013084.OF,013093.OF,013094.OF,013105.OF,013106.OF,013119.OF,013120.OF,013121.OF,013122.OF,013125.OF,013126.OF,013129.OF,013130.OF,013132.OF,013133.OF,013134.OF,013151.OF,013152.OF,013162.OF,013163.OF,013174.OF,013177.OF,013178.OF,013179.OF,013180.OF,013188.OF,013195.OF,013196.OF,013218.OF,013219.OF,013233.OF,013234.OF,013242.OF,013262.OF,013273.OF,013275.OF,013276.OF,013277.OF,013278.OF,013286.OF,013291.OF,013292.OF,013298.OF,013299.OF,013302.OF,013303.OF,013304.OF,013305.OF,013306.OF,013307.OF,013310.OF,013311.OF,013313.OF,013314.OF,013315.OF,013316.OF,013317.OF,013318.OF,013319.OF,013320.OF,013330.OF,013331.OF,013332.OF,013339.OF,013340.OF,013355.OF,013413.OF,013414.OF,013415.OF,013416.OF,013437.OF,013438.OF,013442.OF,013443.OF,013444.OF,013445.OF,013446.OF,013447.OF,013448.OF,013453.OF,013465.OF,013466.OF,013471.OF,013472.OF,013473.OF,013474.OF,013475.OF,013476.OF,013477.OF,013478.OF,013479.OF,013480.OF,013481.OF,013490.OF,013491.OF,013502.OF,013503.OF,013505.OF,013511.OF,013596.OF,013597.OF,013601.OF,013602.OF,013605.OF,013606.OF,013613.OF,013622.OF,013639.OF,013640.OF,013641.OF,013642.OF,013802.OF,013803.OF,013810.OF,013811.OF,013816.OF,013817.OF,013833.OF,013834.OF,013869.OF,013870.OF,013873.OF,013874.OF,013878.OF,013879.OF,013881.OF,013899.OF,013900.OF,013907.OF,013908.OF,013918.OF,013919.OF,013922.OF,013923.OF,013926.OF,013927.OF,013940.OF,013941.OF,013942.OF,013943.OF,014018.OF,014019.OF,014028.OF,014046.OF,014052.OF,014062.OF,014063.OF,014064.OF,014099.OF,014100.OF,014110.OF,014111.OF,014117.OF,014118.OF,014125.OF,014126.OF,014128.OF,014129.OF,014130.OF,014133.OF,014134.OF,014153.OF,014154.OF,014155.OF,014156.OF,014163.OF,014172.OF,014173.OF,014174.OF,014185.OF,014186.OF,014187.OF,014188.OF,014191.OF,014192.OF,014193.OF,014194.OF,014198.OF,014201.OF,014202.OF,014218.OF,014219.OF,014226.OF,014227.OF,014237.OF,014238.OF,014256.OF,014299.OF,014300.OF,014303.OF,014304.OF,014331.OF,014332.OF,014344.OF,014345.OF,014364.OF,014372.OF,014376.OF,014377.OF,014380.OF,014382.OF,014406.OF,014407.OF,014416.OF,014417.OF,014418.OF,014419.OF,014431.OF,014432.OF,014453.OF,014455.OF,014519.OF,014520.OF,014528.OF,014529.OF,014530.OF,014531.OF,014532.OF,014533.OF,014534.OF,014535.OF,014536.OF,014543.OF,014544.OF,014602.OF,014603.OF,014604.OF,014605.OF,014629.OF,014630.OF,014631.OF,014673.OF,014674.OF,014689.OF,014690.OF,014736.OF,014737.OF,014758.OF,014776.OF,014777.OF,014820.OF,014821.OF,014831.OF,014832.OF,014863.OF,014864.OF,014866.OF,014902.OF,014903.OF,014906.OF,014907.OF,014908.OF,014909.OF,014930.OF,014932.OF,014937.OF,014942.OF,014943.OF,014967.OF,014979.OF,014980.OF,014983.OF,014984.OF,014985.OF,014994.OF,014997.OF,015001.OF,015002.OF,015037.OF,015038.OF,015040.OF,015042.OF,015046.OF,015048.OF,015052.OF,015053.OF,015060.OF,015075.OF,015085.OF,015086.OF,015114.OF,015115.OF,015123.OF,015124.OF,015139.OF,015140.OF,015157.OF,015159.OF,015170.OF,015171.OF,015172.OF,015176.OF,015177.OF,015178.OF,015194.OF,015195.OF,015196.OF,015197.OF,015198.OF,015199.OF,015225.OF,015226.OF,015278.OF,015279.OF,015328.OF,015329.OF,015336.OF,015337.OF,015364.OF,015387.OF,015388.OF,015395.OF,015468.OF,015475.OF,015484.OF,020011.OF,020021.OF,040002.OF,040180.OF,040190.OF,050002.OF,050013.OF,050021.OF,050024.OF,070023.OF,070030.OF,070039.OF,090010.OF,090012.OF,100032.OF,100038.OF,100053.OF,110003.OF,110019.OF,110020.OF,110021.OF,110022.OF,110026.OF,110030.OF,159601.OF,159602.OF,159606.OF,159608.OF,159610.OF,159611.OF,159613.OF,159618.OF,159619.OF,159701.OF,159702.OF,159703.OF,159706.OF,159707.OF,159708.OF,159709.OF,159710.OF,159711.OF,159712.OF,159713.OF,159715.OF,159716.OF,159717.OF,159718.OF,159719.OF,159720.OF,159721.OF,159723.OF,159725.OF,159726.OF,159728.OF,159729.OF,159730.OF,159731.OF,159732.OF,159733.OF,159735.OF,159736.OF,159738.OF,159739.OF,159743.OF,159745.OF,159748.OF,159751.OF,159752.OF,159755.OF,159757.OF,159758.OF,159760.OF,159761.OF,159763.OF,159766.OF,159767.OF,159768.OF,159769.OF,159770.OF,159773.OF,159775.OF,159776.OF,159777.OF,159778.OF,159779.OF,159780.OF,159781.OF,159782.OF,159783.OF,159786.OF,159787.OF,159788.OF,159789.OF,159790.OF,159791.OF,159792.OF,159793.OF,159795.OF,159796.OF,159801.OF,159804.OF,159805.OF,159806.OF,159807.OF,159808.OF,159810.OF,159811.OF,159813.OF,159814.OF,159815.OF,159819.OF,159820.OF,159821.OF,159824.OF,159825.OF,159827.OF,159828.OF,159835.OF,159836.OF,159837.OF,159838.OF,159839.OF,159840.OF,159841.OF,159842.OF,159843.OF,159845.OF,159847.OF,159848.OF,159849.OF,159851.OF,159852.OF,159853.OF,159855.OF,159856.OF,159857.OF,159858.OF,159859.OF,159861.OF,159862.OF,159863.OF,159864.OF,159865.OF,159867.OF,159869.OF,159870.OF,159871.OF,159872.OF,159873.OF,159875.OF,159876.OF,159877.OF,159880.OF,159881.OF,159883.OF,159885.OF,159886.OF,159887.OF,159888.OF,159889.OF,159890.OF,159891.OF,159895.OF,159896.OF,159897.OF,159898.OF,159899.OF,159901.OF,159902.OF,159903.OF,159905.OF,159906.OF,159907.OF,159908.OF,159909.OF,159910.OF,159912.OF,159913.OF,159915.OF,159916.OF,159918.OF,159919.OF,159922.OF,159923.OF,159925.OF,159928.OF,159929.OF,159930.OF,159931.OF,159932.OF,159933.OF,159935.OF,159936.OF,159938.OF,159939.OF,159940.OF,159943.OF,159944.OF,159945.OF,159948.OF,159949.OF,159951.OF,159952.OF,159955.OF,159956.OF,159957.OF,159958.OF,159959.OF,159960.OF,159961.OF,159962.OF,159964.OF,159965.OF,159966.OF,159967.OF,159968.OF,159969.OF,159970.OF,159971.OF,159973.OF,159974.OF,159975.OF,159976.OF,159977.OF,159978.OF,159979.OF,159982.OF,159983.OF,159984.OF,159986.OF,159987.OF,159990.OF,159991.OF,159992.OF,159993.OF,159994.OF,159995.OF,159996.OF,159997.OF,159998.OF,159999.OF,160119.OF,160127.OF,160133.OF,160135.OF,160137.OF,160144.OF,160218.OF,160219.OF,160221.OF,160222.OF,160223.OF,160224.OF,160225.OF,160322.OF,160418.OF,160419.OF,160420.OF,160422.OF,160424.OF,160516.OF,160517.OF,160615.OF,160616.OF,160620.OF,160625.OF,160626.OF,160628.OF,160629.OF,160630.OF,160631.OF,160632.OF,160633.OF,160634.OF,160635.OF,160636.OF,160637.OF,160638.OF,160639.OF,160643.OF,160646.OF,160706.OF,160716.OF,160724.OF,160725.OF,160806.OF,160807.OF,160814.OF,160919.OF,160925.OF,161017.OF,161022.OF,161024.OF,161025.OF,161026.OF,161027.OF,161028.OF,161029.OF,161030.OF,161031.OF,161032.OF,161033.OF,161035.OF,161036.OF,161037.OF,161039.OF,161118.OF,161121.OF,161122.OF,161123.OF,161211.OF,161217.OF,161227.OF,161604.OF,161607.OF,161612.OF,161613.OF,161628.OF,161631.OF,161715.OF,161718.OF,161720.OF,161721.OF,161723.OF,161724.OF,161725.OF,161726.OF,161811.OF,161812.OF,161816.OF,161907.OF,162107.OF,162208.OF,162213.OF,162216.OF,162307.OF,162412.OF,162413.OF,162416.OF,162509.OF,162510.OF,162711.OF,162714.OF,162907.OF,163109.OF,163110.OF,163111.OF,163113.OF,163114.OF,163115.OF,163116.OF,163118.OF,163209.OF,163407.OF,163808.OF,163821.OF,164205.OF,164304.OF,164401.OF,164402.OF,164508.OF,164809.OF,164811.OF,164818.OF,164825.OF,164905.OF,164908.OF,165309.OF,165310.OF,165312.OF,165511.OF,165515.OF,165519.OF,165520.OF,165521.OF,165522.OF,165523.OF,165524.OF,165525.OF,165806.OF,165810.OF,166402.OF,166802.OF,167301.OF,167302.OF,167503.OF,167506.OF,167507.OF,167601.OF,167702.OF,167703.OF,168203.OF,168204.OF,168205.OF,168701.OF,168702.OF,180003.OF,200002.OF,202015.OF,202017.OF,202021.OF,202025.OF,202211.OF,206012.OF,213010.OF,217016.OF,217017.OF,217019.OF,217027.OF,233010.OF,240014.OF,240016.OF,240019.OF,257060.OF,270010.OF,270026.OF,290010.OF,310318.OF,310398.OF,320010.OF,320014.OF,320020.OF,320022.OF,340006.OF,360001.OF,370023.OF,376510.OF,399001.OF,399011.OF,410008.OF,410010.OF,450008.OF,450009.OF,460220.OF,460300.OF,470007.OF,470068.OF,481009.OF,481012.OF,501005.OF,501006.OF,501007.OF,501008.OF,501009.OF,501010.OF,501011.OF,501012.OF,501016.OF,501019.OF,501025.OF,501029.OF,501030.OF,501031.OF,501036.OF,501037.OF,501043.OF,501045.OF,501047.OF,501048.OF,501050.OF,501057.OF,501058.OF,501059.OF,501060.OF,501061.OF,501067.OF,501068.OF,501069.OF,501086.OF,501089.OF,501090.OF,501219.OF,501301.OF,501302.OF,501303.OF,501305.OF,501306.OF,501307.OF,501308.OF,501309.OF,501310.OF,501311.OF,502000.OF,502003.OF,502006.OF,502010.OF,502013.OF,502020.OF,502023.OF,502040.OF,502048.OF,502053.OF,502056.OF,510010.OF,510020.OF,510030.OF,510050.OF,510060.OF,510090.OF,510100.OF,510110.OF,510120.OF,510130.OF,510150.OF,510160.OF,510170.OF,510180.OF,510190.OF,510200.OF,510210.OF,510220.OF,510230.OF,510270.OF,510290.OF,510300.OF,510310.OF,510330.OF,510350.OF,510360.OF,510370.OF,510380.OF,510390.OF,510410.OF,510440.OF,510500.OF,510510.OF,510530.OF,510550.OF,510560.OF,510570.OF,510580.OF,510590.OF,510600.OF,510630.OF,510650.OF,510660.OF,510680.OF,510710.OF,510760.OF,510770.OF,510800.OF,510810.OF,510850.OF,510880.OF,510990.OF,512000.OF,512010.OF,512040.OF,512070.OF,512090.OF,512100.OF,512120.OF,512150.OF,512160.OF,512170.OF,512180.OF,512190.OF,512200.OF,512220.OF,512260.OF,512280.OF,512290.OF,512320.OF,512330.OF,512360.OF,512380.OF,512390.OF,512400.OF,512480.OF,512500.OF,512510.OF,512520.OF,512530.OF,512550.OF,512560.OF,512570.OF,512580.OF,512590.OF,512600.OF,512610.OF,512640.OF,512650.OF,512660.OF,512670.OF,512680.OF,512690.OF,512700.OF,512710.OF,512720.OF,512730.OF,512750.OF,512760.OF,512770.OF,512780.OF,512800.OF,512810.OF,512820.OF,512870.OF,512880.OF,512890.OF,512900.OF,512910.OF,512920.OF,512930.OF,512950.OF,512960.OF,512970.OF,512980.OF,512990.OF,513020.OF,513070.OF,513090.OF,513150.OF,513160.OF,513200.OF,513230.OF,513320.OF,513550.OF,513590.OF,513660.OF,513680.OF,513690.OF,513700.OF,513770.OF,513860.OF,513900.OF,513960.OF,513980.OF,513990.OF,515000.OF,515010.OF,515020.OF,515030.OF,515050.OF,515060.OF,515070.OF,515080.OF,515090.OF,515100.OF,515110.OF,515120.OF,515130.OF,515150.OF,515160.OF,515170.OF,515180.OF,515190.OF,515200.OF,515210.OF,515220.OF,515230.OF,515250.OF,515260.OF,515280.OF,515290.OF,515300.OF,515310.OF,515320.OF,515330.OF,515350.OF,515360.OF,515380.OF,515390.OF,515400.OF,515450.OF,515500.OF,515510.OF,515520.OF,515530.OF,515550.OF,515560.OF,515570.OF,515580.OF,515590.OF,515600.OF,515610.OF,515630.OF,515650.OF,515660.OF,515670.OF,515680.OF,515700.OF,515710.OF,515750.OF,515760.OF,515770.OF,515780.OF,515790.OF,515800.OF,515810.OF,515820.OF,515850.OF,515860.OF,515870.OF,515880.OF,515890.OF,515900.OF,515910.OF,515920.OF,515930.OF,515950.OF,515960.OF,515980.OF,515990.OF,516000.OF,516010.OF,516020.OF,516050.OF,516060.OF,516070.OF,516080.OF,516090.OF,516100.OF,516110.OF,516120.OF,516130.OF,516150.OF,516160.OF,516180.OF,516190.OF,516200.OF,516210.OF,516220.OF,516260.OF,516270.OF,516290.OF,516300.OF,516310.OF,516320.OF,516330.OF,516350.OF,516360.OF,516380.OF,516390.OF,516480.OF,516500.OF,516510.OF,516520.OF,516530.OF,516550.OF,516560.OF,516570.OF,516580.OF,516590.OF,516600.OF,516610.OF,516620.OF,516630.OF,516640.OF,516650.OF,516660.OF,516670.OF,516680.OF,516690.OF,516700.OF,516710.OF,516720.OF,516730.OF,516750.OF,516760.OF,516770.OF,516780.OF,516790.OF,516800.OF,516810.OF,516820.OF,516830.OF,516850.OF,516860.OF,516880.OF,516890.OF,516900.OF,516910.OF,516920.OF,516930.OF,516950.OF,516960.OF,516970.OF,516980.OF,517000.OF,517010.OF,517030.OF,517050.OF,517080.OF,517090.OF,517100.OF,517110.OF,517120.OF,517160.OF,517170.OF,517180.OF,517200.OF,517270.OF,517280.OF,517300.OF,517330.OF,517350.OF,517360.OF,517380.OF,517390.OF,517500.OF,517550.OF,517660.OF,517760.OF,517770.OF,517780.OF,517800.OF,517880.OF,517900.OF,517960.OF,519027.OF,519032.OF,519034.OF,519100.OF,519116.OF,519117.OF,519180.OF,519193.OF,519300.OF,519606.OF,519671.OF,519673.OF,519677.OF,519686.OF,519706.OF,519714.OF,519935.OF,519965.OF,519975.OF,530010.OF,530015.OF,530018.OF,540006.OF,540007.OF,540008.OF,540009.OF,540010.OF,540012.OF,560000.OF,560050.OF,560100.OF,560500.OF,560560.OF,560600.OF,560660.OF,560800.OF,560880.OF,561100.OF,561120.OF,561130.OF,561300.OF,561310.OF,561350.OF,561500.OF,561550.OF,561600.OF,561800.OF,561900.OF,561910.OF,561990.OF,562300.OF,562360.OF,562500.OF,562510.OF,562520.OF,562800.OF,562880.OF,562900.OF,562910.OF,562950.OF,563000.OF,580008.OF,585001.OF,588000.OF,588050.OF,588060.OF,588080.OF,588090.OF,588150.OF,588180.OF,588280.OF,588300.OF,588310.OF,588330.OF,588360.OF,588380.OF,588390.OF,588400.OF,590007.OF,660008.OF,660011.OF,671030.OF,673090.OF,673100.OF,673101.OF,690008.OF,700002.OF,740101.OF,900029.OF,900030.OF,920002.OF,920003.OF,920922.OF,920923.OF,960000.OF,960022.OF,970041.OF,970042.OF,H161604.OF,H202017.OF','ths_fund_scale_fund','2017-04-01')
        fund_list=all_fund.data.dropna()["thscode"].to_list()
        for i in fund_list:
            data_result = THS_HQ(i,'netAssetValue','Interval:M','2017-03-01','2022-03-01')
            if data_result.errorcode != 0:
                print('error:{}'.format(data_result.errmsg))
            else:
                data_df=data_result.data
                data_df.to_csv('{}.csv'.format(i))
    return data_df



def datapool_bd():
    data_fund=THS_BD('000008.OF,000042.OF,000051.OF,000059.OF,000082.OF,000176.OF,000248.OF,000309.OF,000311.OF,000312.OF,000313.OF,000326.OF,000368.OF,000373.OF,000376.OF,000409.OF,000411.OF,000418.OF,000457.OF,000471.OF,000478.OF,000512.OF,000513.OF,000524.OF,000549.OF,000577.OF,000586.OF,000592.OF,000594.OF,000596.OF,000613.OF,000628.OF,000656.OF,000688.OF,000696.OF,000697.OF,000711.OF,000729.OF,000746.OF,000751.OF,000756.OF,000761.OF,000778.OF,000780.OF,000793.OF,000803.OF,000826.OF,000827.OF,000828.OF,000831.OF,000835.OF,000854.OF,000866.OF,000867.OF,000884.OF,000893.OF,000913.OF,000916.OF,000925.OF,000942.OF,000948.OF,000950.OF,000955.OF,000960.OF,000961.OF,000962.OF,000968.OF,000971.OF,000974.OF,000975.OF,000978.OF,000979.OF,000985.OF,000991.OF,000996.OF,001008.OF,001009.OF,001015.OF,001016.OF,001027.OF,001028.OF,001036.OF,001039.OF,001040.OF,001042.OF,001043.OF,001044.OF,001047.OF,001048.OF,001050.OF,001051.OF,001052.OF,001054.OF,001064.OF,001070.OF,001072.OF,001097.OF,001104.OF,001105.OF,001113.OF,001126.OF,001133.OF,001149.OF,001158.OF,001162.OF,001163.OF,001166.OF,001167.OF,001171.OF,001178.OF,001180.OF,001186.OF,001188.OF,001193.OF,001195.OF,001208.OF,001214.OF,001223.OF,001230.OF,001236.OF,001237.OF,001241.OF,001242.OF,001243.OF,001245.OF,001277.OF,001291.OF,001313.OF,001319.OF,001344.OF,001351.OF,001361.OF,001396.OF,001397.OF,001404.OF,001409.OF,001410.OF,001416.OF,001420.OF,001421.OF,001426.OF,001455.OF,001469.OF,001473.OF,001476.OF,001482.OF,001490.OF,001496.OF,001520.OF,001521.OF,001528.OF,001539.OF,001541.OF,001542.OF,001548.OF,001549.OF,001550.OF,001551.OF,001552.OF,001553.OF,001556.OF,001557.OF,001576.OF,001577.OF,001579.OF,001583.OF,001588.OF,001589.OF,001592.OF,001593.OF,001594.OF,001595.OF,001605.OF,001616.OF,001617.OF,001618.OF,001626.OF,001628.OF,001629.OF,001630.OF,001631.OF,001632.OF,001637.OF,001638.OF,001643.OF,001644.OF,001645.OF,001651.OF,001672.OF,001677.OF,001678.OF,001685.OF,001692.OF,001703.OF,001705.OF,001707.OF,001714.OF,001717.OF,001718.OF,001719.OF,001725.OF,001726.OF,001733.OF,001736.OF,001749.OF,001764.OF,001766.OF,001781.OF,001790.OF,001825.OF,001849.OF,001877.OF,001878.OF,001879.OF,001899.OF,001915.OF,001917.OF,001938.OF,001956.OF,001974.OF,001975.OF,002063.OF,002076.OF,002121.OF,002168.OF,002199.OF,002210.OF,002229.OF,002236.OF,002289.OF,002300.OF,002310.OF,002311.OF,002315.OF,002316.OF,002332.OF,002333.OF,002334.OF,002335.OF,002385.OF,002387.OF,002510.OF,002556.OF,002595.OF,002621.OF,002656.OF,002670.OF,002671.OF,002697.OF,002861.OF,002871.OF,002872.OF,002900.OF,002903.OF,002906.OF,002907.OF,002952.OF,002974.OF,002977.OF,002978.OF,002979.OF,002980.OF,002982.OF,002984.OF,002987.OF,003015.OF,003016.OF,003017.OF,003053.OF,003054.OF,003069.OF,003070.OF,003145.OF,003184.OF,003190.OF,003191.OF,003194.OF,003230.OF,003231.OF,003232.OF,003233.OF,003298.OF,003299.OF,003318.OF,003359.OF,003366.OF,003416.OF,003475.OF,003548.OF,003578.OF,003579.OF,003624.OF,003625.OF,003634.OF,003646.OF,003647.OF,003745.OF,003760.OF,003761.OF,003765.OF,003766.OF,003834.OF,003853.OF,003865.OF,003876.OF,003884.OF,003885.OF,003956.OF,003957.OF,003958.OF,003984.OF,003985.OF,003986.OF,004040.OF,004041.OF,004069.OF,004070.OF,004075.OF,004098.OF,004189.OF,004190.OF,004191.OF,004192.OF,004193.OF,004194.OF,004195.OF,004241.OF,004292.OF,004342.OF,004343.OF,004344.OF,004345.OF,004346.OF,004347.OF,004348.OF,004350.OF,004352.OF,004403.OF,004404.OF,004407.OF,004408.OF,004409.OF,004410.OF,004432.OF,004433.OF,004450.OF,004476.OF,004488.OF,004512.OF,004513.OF,004532.OF,004533.OF,004597.OF,004598.OF,004606.OF,004616.OF,004640.OF,004642.OF,004643.OF,004683.OF,004686.OF,004698.OF,004716.OF,004730.OF,004742.OF,004743.OF,004744.OF,004746.OF,004752.OF,004753.OF,004784.OF,004788.OF,004789.OF,004790.OF,004791.OF,004805.OF,004812.OF,004813.OF,004851.OF,004854.OF,004855.OF,004856.OF,004857.OF,004858.OF,004870.OF,004874.OF,004875.OF,004876.OF,004891.OF,004892.OF,004925.OF,004945.OF,004995.OF,004996.OF,004997.OF,005009.OF,005035.OF,005036.OF,005037.OF,005038.OF,005051.OF,005052.OF,005062.OF,005063.OF,005064.OF,005102.OF,005103.OF,005106.OF,005112.OF,005113.OF,005114.OF,005125.OF,005137.OF,005152.OF,005161.OF,005188.OF,005189.OF,005209.OF,005210.OF,005223.OF,005224.OF,005229.OF,005235.OF,005236.OF,005237.OF,005238.OF,005248.OF,005259.OF,005267.OF,005268.OF,005287.OF,005288.OF,005303.OF,005304.OF,005310.OF,005313.OF,005314.OF,005328.OF,005390.OF,005391.OF,005402.OF,005433.OF,005457.OF,005495.OF,005496.OF,005502.OF,005505.OF,005506.OF,005530.OF,005554.OF,005555.OF,005561.OF,005562.OF,005569.OF,005570.OF,005607.OF,005608.OF,005612.OF,005620.OF,005621.OF,005626.OF,005627.OF,005628.OF,005629.OF,005633.OF,005635.OF,005636.OF,005639.OF,005640.OF,005658.OF,005659.OF,005660.OF,005661.OF,005662.OF,005663.OF,005669.OF,005691.OF,005693.OF,005702.OF,005707.OF,005727.OF,005733.OF,005734.OF,005735.OF,005737.OF,005761.OF,005762.OF,005763.OF,005770.OF,005777.OF,005788.OF,005789.OF,005795.OF,005802.OF,005813.OF,005814.OF,005825.OF,005826.OF,005829.OF,005830.OF,005832.OF,005867.OF,005868.OF,005869.OF,005870.OF,005873.OF,005874.OF,005880.OF,005881.OF,005885.OF,005894.OF,005918.OF,005919.OF,005927.OF,005928.OF,005960.OF,005961.OF,005962.OF,005963.OF,005965.OF,005966.OF,005968.OF,005969.OF,005970.OF,005983.OF,005994.OF,005998.OF,005999.OF,006002.OF,006003.OF,006020.OF,006021.OF,006034.OF,006048.OF,006063.OF,006080.OF,006081.OF,006087.OF,006098.OF,006106.OF,006119.OF,006129.OF,006130.OF,006131.OF,006138.OF,006143.OF,006144.OF,006165.OF,006166.OF,006193.OF,006194.OF,006195.OF,006196.OF,006197.OF,006214.OF,006215.OF,006220.OF,006221.OF,006228.OF,006229.OF,006248.OF,006249.OF,006251.OF,006257.OF,006265.OF,006286.OF,006293.OF,006339.OF,006341.OF,006342.OF,006346.OF,006347.OF,006355.OF,006363.OF,006382.OF,006395.OF,006438.OF,006439.OF,006440.OF,006441.OF,006478.OF,006486.OF,006487.OF,006502.OF,006503.OF,006524.OF,006525.OF,006560.OF,006561.OF,006569.OF,006593.OF,006594.OF,006600.OF,006603.OF,006604.OF,006605.OF,006611.OF,006614.OF,006656.OF,006657.OF,006658.OF,006659.OF,006671.OF,006675.OF,006676.OF,006682.OF,006687.OF,006688.OF,006692.OF,006693.OF,006697.OF,006704.OF,006705.OF,006712.OF,006713.OF,006724.OF,006729.OF,006730.OF,006733.OF,006743.OF,006744.OF,006748.OF,006751.OF,006756.OF,006757.OF,006781.OF,006783.OF,006784.OF,006786.OF,006787.OF,006803.OF,006809.OF,006810.OF,006818.OF,006819.OF,006868.OF,006909.OF,006910.OF,006911.OF,006912.OF,006923.OF,006924.OF,006926.OF,006928.OF,006937.OF,006938.OF,006939.OF,006957.OF,006981.OF,007005.OF,007028.OF,007029.OF,007039.OF,007044.OF,007045.OF,007076.OF,007077.OF,007089.OF,007096.OF,007107.OF,007108.OF,007130.OF,007133.OF,007134.OF,007135.OF,007136.OF,007137.OF,007138.OF,007143.OF,007144.OF,007153.OF,007154.OF,007178.OF,007191.OF,007216.OF,007223.OF,007230.OF,007275.OF,007276.OF,007287.OF,007288.OF,007300.OF,007301.OF,007339.OF,007354.OF,007357.OF,007379.OF,007380.OF,007386.OF,007388.OF,007389.OF,007397.OF,007404.OF,007405.OF,007410.OF,007411.OF,007413.OF,007431.OF,007448.OF,007464.OF,007465.OF,007466.OF,007467.OF,007470.OF,007471.OF,007472.OF,007473.OF,007474.OF,007475.OF,007505.OF,007506.OF,007512.OF,007523.OF,007524.OF,007531.OF,007538.OF,007539.OF,007541.OF,007548.OF,007571.OF,007580.OF,007586.OF,007590.OF,007593.OF,007594.OF,007605.OF,007606.OF,007657.OF,007658.OF,007664.OF,007665.OF,007671.OF,007672.OF,007674.OF,007675.OF,007685.OF,007737.OF,007750.OF,007751.OF,007760.OF,007782.OF,007783.OF,007784.OF,007785.OF,007786.OF,007787.OF,007788.OF,007789.OF,007792.OF,007793.OF,007794.OF,007795.OF,007796.OF,007797.OF,007799.OF,007800.OF,007801.OF,007804.OF,007806.OF,007807.OF,007809.OF,007810.OF,007815.OF,007816.OF,007817.OF,007818.OF,007831.OF,007832.OF,007839.OF,007840.OF,007842.OF,007843.OF,007853.OF,007856.OF,007857.OF,007873.OF,007874.OF,007882.OF,007883.OF,007885.OF,007886.OF,007903.OF,007932.OF,007943.OF,007950.OF,007952.OF,007965.OF,007966.OF,007983.OF,007984.OF,007992.OF,007993.OF,007994.OF,007995.OF,008001.OF,008009.OF,008020.OF,008021.OF,008052.OF,008053.OF,008056.OF,008057.OF,008072.OF,008075.OF,008076.OF,008084.OF,008085.OF,008086.OF,008087.OF,008088.OF,008089.OF,008107.OF,008112.OF,008113.OF,008114.OF,008115.OF,008124.OF,008128.OF,008129.OF,008134.OF,008154.OF,008155.OF,008163.OF,008164.OF,008166.OF,008167.OF,008177.OF,008184.OF,008189.OF,008190.OF,008199.OF,008200.OF,008201.OF,008234.OF,008236.OF,008237.OF,008238.OF,008239.OF,008240.OF,008241.OF,008249.OF,008250.OF,008258.OF,008259.OF,008261.OF,008262.OF,008264.OF,008265.OF,008279.OF,008280.OF,008281.OF,008282.OF,008283.OF,008286.OF,008291.OF,008292.OF,008294.OF,008295.OF,008298.OF,008299.OF,008303.OF,008304.OF,008314.OF,008315.OF,008326.OF,008327.OF,008382.OF,008390.OF,008391.OF,008396.OF,008397.OF,008399.OF,008400.OF,008407.OF,008408.OF,008412.OF,008413.OF,008445.OF,008519.OF,008520.OF,008537.OF,008538.OF,008551.OF,008552.OF,008585.OF,008586.OF,008590.OF,008591.OF,008592.OF,008593.OF,008618.OF,008619.OF,008682.OF,008709.OF,008713.OF,008714.OF,008736.OF,008737.OF,008749.OF,008750.OF,008768.OF,008769.OF,008776.OF,008777.OF,008778.OF,008779.OF,008832.OF,008854.OF,008855.OF,008860.OF,008871.OF,008872.OF,008887.OF,008888.OF,008889.OF,008898.OF,008899.OF,008907.OF,008908.OF,008916.OF,008917.OF,008923.OF,008924.OF,008926.OF,008927.OF,008928.OF,008929.OF,008934.OF,008935.OF,008944.OF,008945.OF,008958.OF,008973.OF,008975.OF,008976.OF,009004.OF,009012.OF,009013.OF,009017.OF,009043.OF,009046.OF,009047.OF,009051.OF,009052.OF,009059.OF,009060.OF,009067.OF,009068.OF,009069.OF,009070.OF,009079.OF,009080.OF,009106.OF,009107.OF,009116.OF,009117.OF,009126.OF,009127.OF,009147.OF,009162.OF,009163.OF,009179.OF,009180.OF,009194.OF,009195.OF,009208.OF,009239.OF,009240.OF,009265.OF,009300.OF,009329.OF,009336.OF,009337.OF,009341.OF,009374.OF,009375.OF,009379.OF,009384.OF,009398.OF,009414.OF,009439.OF,009472.OF,009476.OF,009479.OF,009480.OF,009491.OF,009492.OF,009502.OF,009503.OF,009513.OF,009514.OF,009537.OF,009538.OF,009550.OF,009551.OF,009557.OF,009601.OF,009602.OF,009608.OF,009609.OF,009613.OF,009614.OF,009658.OF,009726.OF,009727.OF,009733.OF,009734.OF,009775.OF,009794.OF,009805.OF,009852.OF,009860.OF,009863.OF,009864.OF,009865.OF,009874.OF,009875.OF,009876.OF,009877.OF,009881.OF,009891.OF,009896.OF,009897.OF,009898.OF,009899.OF,009904.OF,009905.OF,009954.OF,009981.OF,009982.OF,009988.OF,010001.OF,010002.OF,010003.OF,010004.OF,010010.OF,010013.OF,010024.OF,010066.OF,010067.OF,010112.OF,010113.OF,010120.OF,010121.OF,010128.OF,010129.OF,010141.OF,010142.OF,010144.OF,010153.OF,010154.OF,010157.OF,010158.OF,010160.OF,010161.OF,010162.OF,010183.OF,010196.OF,010197.OF,010202.OF,010203.OF,010204.OF,010210.OF,010224.OF,010235.OF,010236.OF,010245.OF,010253.OF,010321.OF,010322.OF,010331.OF,010332.OF,010341.OF,010342.OF,010349.OF,010351.OF,010352.OF,010355.OF,010356.OF,010361.OF,010362.OF,010364.OF,010365.OF,010366.OF,010387.OF,010388.OF,010391.OF,010392.OF,010409.OF,010419.OF,010420.OF,010432.OF,010434.OF,010469.OF,010470.OF,010495.OF,010496.OF,010505.OF,010524.OF,010531.OF,010556.OF,010561.OF,010572.OF,010573.OF,010585.OF,010586.OF,010592.OF,010593.OF,010673.OF,010674.OF,010677.OF,010680.OF,010681.OF,010685.OF,010687.OF,010703.OF,010704.OF,010709.OF,010710.OF,010736.OF,010737.OF,010749.OF,010750.OF,010769.OF,010770.OF,010771.OF,010772.OF,010777.OF,010778.OF,010785.OF,010786.OF,010801.OF,010802.OF,010805.OF,010806.OF,010812.OF,010815.OF,010854.OF,010855.OF,010872.OF,010873.OF,010892.OF,010893.OF,010896.OF,010897.OF,010908.OF,010909.OF,010953.OF,010954.OF,010955.OF,010956.OF,010957.OF,010961.OF,010989.OF,010990.OF,010992.OF,010993.OF,010998.OF,011013.OF,011014.OF,011035.OF,011036.OF,011040.OF,011041.OF,011042.OF,011043.OF,011066.OF,011077.OF,011102.OF,011103.OF,011104.OF,011107.OF,011108.OF,011122.OF,011123.OF,011125.OF,011126.OF,011132.OF,011133.OF,011142.OF,011143.OF,011146.OF,011147.OF,011149.OF,011150.OF,011179.OF,011180.OF,011205.OF,011220.OF,011221.OF,011224.OF,011225.OF,011229.OF,011230.OF,011236.OF,011237.OF,011254.OF,011255.OF,011269.OF,011270.OF,011285.OF,011286.OF,011295.OF,011316.OF,011317.OF,011319.OF,011320.OF,011321.OF,011322.OF,011323.OF,011326.OF,011328.OF,011329.OF,011349.OF,011350.OF,011373.OF,011374.OF,011377.OF,011378.OF,011384.OF,011385.OF,011418.OF,011419.OF,011424.OF,011425.OF,011426.OF,011457.OF,011463.OF,011470.OF,011473.OF,011476.OF,011478.OF,011498.OF,011499.OF,011506.OF,011507.OF,011512.OF,011513.OF,011545.OF,011546.OF,011589.OF,011590.OF,011592.OF,011601.OF,011602.OF,011607.OF,011608.OF,011609.OF,011610.OF,011611.OF,011612.OF,011613.OF,011614.OF,011615.OF,011630.OF,011631.OF,011645.OF,011646.OF,011651.OF,011652.OF,011667.OF,011668.OF,011685.OF,011686.OF,011722.OF,011723.OF,011726.OF,011793.OF,011824.OF,011825.OF,011832.OF,011833.OF,011836.OF,011839.OF,011840.OF,011853.OF,011854.OF,011860.OF,011861.OF,011882.OF,011883.OF,011923.OF,011926.OF,011934.OF,011935.OF,011966.OF,011967.OF,011971.OF,011972.OF,012040.OF,012041.OF,012042.OF,012043.OF,012044.OF,012045.OF,012046.OF,012080.OF,012081.OF,012116.OF,012117.OF,012126.OF,012127.OF,012151.OF,012152.OF,012157.OF,012158.OF,012179.OF,012180.OF,012181.OF,012206.OF,012207.OF,012210.OF,012211.OF,012212.OF,012213.OF,012214.OF,012215.OF,012238.OF,012244.OF,012245.OF,012275.OF,012276.OF,012315.OF,012316.OF,012319.OF,012320.OF,012321.OF,012322.OF,012323.OF,012326.OF,012327.OF,012340.OF,012341.OF,012362.OF,012363.OF,012364.OF,012365.OF,012371.OF,012372.OF,012394.OF,012401.OF,012402.OF,012405.OF,012414.OF,012415.OF,012416.OF,012417.OF,012419.OF,012445.OF,012461.OF,012462.OF,012496.OF,012497.OF,012498.OF,012499.OF,012503.OF,012504.OF,012516.OF,012517.OF,012537.OF,012538.OF,012543.OF,012544.OF,012547.OF,012548.OF,012549.OF,012550.OF,012551.OF,012552.OF,012553.OF,012554.OF,012555.OF,012559.OF,012560.OF,012561.OF,012562.OF,012590.OF,012596.OF,012597.OF,012598.OF,012599.OF,012600.OF,012605.OF,012606.OF,012619.OF,012620.OF,012629.OF,012630.OF,012634.OF,012635.OF,012636.OF,012637.OF,012643.OF,012644.OF,012645.OF,012646.OF,012655.OF,012658.OF,012663.OF,012664.OF,012679.OF,012680.OF,012694.OF,012695.OF,012696.OF,012697.OF,012698.OF,012699.OF,012700.OF,012708.OF,012709.OF,012712.OF,012713.OF,012717.OF,012718.OF,012722.OF,012723.OF,012724.OF,012725.OF,012728.OF,012729.OF,012730.OF,012731.OF,012733.OF,012734.OF,012737.OF,012738.OF,012754.OF,012755.OF,012756.OF,012757.OF,012759.OF,012760.OF,012761.OF,012762.OF,012763.OF,012764.OF,012768.OF,012769.OF,012781.OF,012782.OF,012800.OF,012801.OF,012802.OF,012808.OF,012809.OF,012810.OF,012811.OF,012831.OF,012832.OF,012835.OF,012836.OF,012837.OF,012838.OF,012842.OF,012854.OF,012855.OF,012857.OF,012862.OF,012863.OF,012872.OF,012873.OF,012874.OF,012875.OF,012882.OF,012883.OF,012884.OF,012885.OF,012886.OF,012894.OF,012895.OF,012898.OF,012899.OF,012900.OF,012901.OF,012907.OF,012908.OF,012911.OF,012912.OF,012926.OF,012927.OF,012930.OF,012961.OF,012962.OF,012969.OF,012970.OF,012973.OF,012974.OF,012989.OF,012990.OF,013004.OF,013005.OF,013011.OF,013012.OF,013013.OF,013014.OF,013019.OF,013020.OF,013031.OF,013032.OF,013035.OF,013048.OF,013049.OF,013050.OF,013053.OF,013054.OF,013074.OF,013080.OF,013081.OF,013082.OF,013083.OF,013084.OF,013093.OF,013094.OF,013105.OF,013106.OF,013119.OF,013120.OF,013121.OF,013122.OF,013125.OF,013126.OF,013129.OF,013130.OF,013132.OF,013133.OF,013134.OF,013151.OF,013152.OF,013162.OF,013163.OF,013174.OF,013177.OF,013178.OF,013179.OF,013180.OF,013188.OF,013195.OF,013196.OF,013218.OF,013219.OF,013233.OF,013234.OF,013242.OF,013262.OF,013273.OF,013275.OF,013276.OF,013277.OF,013278.OF,013286.OF,013291.OF,013292.OF,013298.OF,013299.OF,013302.OF,013303.OF,013304.OF,013305.OF,013306.OF,013307.OF,013310.OF,013311.OF,013313.OF,013314.OF,013315.OF,013316.OF,013317.OF,013318.OF,013319.OF,013320.OF,013330.OF,013331.OF,013332.OF,013339.OF,013340.OF,013355.OF,013413.OF,013414.OF,013415.OF,013416.OF,013437.OF,013438.OF,013442.OF,013443.OF,013444.OF,013445.OF,013446.OF,013447.OF,013448.OF,013453.OF,013465.OF,013466.OF,013471.OF,013472.OF,013473.OF,013474.OF,013475.OF,013476.OF,013477.OF,013478.OF,013479.OF,013480.OF,013481.OF,013490.OF,013491.OF,013502.OF,013503.OF,013505.OF,013511.OF,013596.OF,013597.OF,013601.OF,013602.OF,013605.OF,013606.OF,013613.OF,013622.OF,013639.OF,013640.OF,013641.OF,013642.OF,013802.OF,013803.OF,013810.OF,013811.OF,013816.OF,013817.OF,013833.OF,013834.OF,013869.OF,013870.OF,013873.OF,013874.OF,013878.OF,013879.OF,013881.OF,013899.OF,013900.OF,013907.OF,013908.OF,013918.OF,013919.OF,013922.OF,013923.OF,013926.OF,013927.OF,013940.OF,013941.OF,013942.OF,013943.OF,014018.OF,014019.OF,014028.OF,014046.OF,014052.OF,014062.OF,014063.OF,014064.OF,014099.OF,014100.OF,014110.OF,014111.OF,014117.OF,014118.OF,014125.OF,014126.OF,014128.OF,014129.OF,014130.OF,014133.OF,014134.OF,014153.OF,014154.OF,014155.OF,014156.OF,014163.OF,014172.OF,014173.OF,014174.OF,014185.OF,014186.OF,014187.OF,014188.OF,014191.OF,014192.OF,014193.OF,014194.OF,014198.OF,014201.OF,014202.OF,014218.OF,014219.OF,014226.OF,014227.OF,014237.OF,014238.OF,014256.OF,014299.OF,014300.OF,014303.OF,014304.OF,014331.OF,014332.OF,014344.OF,014345.OF,014364.OF,014372.OF,014376.OF,014377.OF,014380.OF,014382.OF,014406.OF,014407.OF,014416.OF,014417.OF,014418.OF,014419.OF,014431.OF,014432.OF,014453.OF,014455.OF,014519.OF,014520.OF,014528.OF,014529.OF,014530.OF,014531.OF,014532.OF,014533.OF,014534.OF,014535.OF,014536.OF,014543.OF,014544.OF,014602.OF,014603.OF,014604.OF,014605.OF,014629.OF,014630.OF,014631.OF,014673.OF,014674.OF,014689.OF,014690.OF,014736.OF,014737.OF,014758.OF,014776.OF,014777.OF,014820.OF,014821.OF,014831.OF,014832.OF,014863.OF,014864.OF,014866.OF,014902.OF,014903.OF,014906.OF,014907.OF,014908.OF,014909.OF,014930.OF,014932.OF,014937.OF,014942.OF,014943.OF,014967.OF,014979.OF,014980.OF,014983.OF,014984.OF,014985.OF,014994.OF,014997.OF,015001.OF,015002.OF,015037.OF,015038.OF,015040.OF,015042.OF,015046.OF,015048.OF,015052.OF,015053.OF,015060.OF,015075.OF,015085.OF,015086.OF,015114.OF,015115.OF,015123.OF,015124.OF,015139.OF,015140.OF,015157.OF,015159.OF,015170.OF,015171.OF,015172.OF,015176.OF,015177.OF,015178.OF,015194.OF,015195.OF,015196.OF,015197.OF,015198.OF,015199.OF,015225.OF,015226.OF,015278.OF,015279.OF,015328.OF,015329.OF,015336.OF,015337.OF,015364.OF,015387.OF,015388.OF,015395.OF,015468.OF,015475.OF,015484.OF,020011.OF,020021.OF,040002.OF,040180.OF,040190.OF,050002.OF,050013.OF,050021.OF,050024.OF,070023.OF,070030.OF,070039.OF,090010.OF,090012.OF,100032.OF,100038.OF,100053.OF,110003.OF,110019.OF,110020.OF,110021.OF,110022.OF,110026.OF,110030.OF,159601.OF,159602.OF,159606.OF,159608.OF,159610.OF,159611.OF,159613.OF,159618.OF,159619.OF,159701.OF,159702.OF,159703.OF,159706.OF,159707.OF,159708.OF,159709.OF,159710.OF,159711.OF,159712.OF,159713.OF,159715.OF,159716.OF,159717.OF,159718.OF,159719.OF,159720.OF,159721.OF,159723.OF,159725.OF,159726.OF,159728.OF,159729.OF,159730.OF,159731.OF,159732.OF,159733.OF,159735.OF,159736.OF,159738.OF,159739.OF,159743.OF,159745.OF,159748.OF,159751.OF,159752.OF,159755.OF,159757.OF,159758.OF,159760.OF,159761.OF,159763.OF,159766.OF,159767.OF,159768.OF,159769.OF,159770.OF,159773.OF,159775.OF,159776.OF,159777.OF,159778.OF,159779.OF,159780.OF,159781.OF,159782.OF,159783.OF,159786.OF,159787.OF,159788.OF,159789.OF,159790.OF,159791.OF,159792.OF,159793.OF,159795.OF,159796.OF,159801.OF,159804.OF,159805.OF,159806.OF,159807.OF,159808.OF,159810.OF,159811.OF,159813.OF,159814.OF,159815.OF,159819.OF,159820.OF,159821.OF,159824.OF,159825.OF,159827.OF,159828.OF,159835.OF,159836.OF,159837.OF,159838.OF,159839.OF,159840.OF,159841.OF,159842.OF,159843.OF,159845.OF,159847.OF,159848.OF,159849.OF,159851.OF,159852.OF,159853.OF,159855.OF,159856.OF,159857.OF,159858.OF,159859.OF,159861.OF,159862.OF,159863.OF,159864.OF,159865.OF,159867.OF,159869.OF,159870.OF,159871.OF,159872.OF,159873.OF,159875.OF,159876.OF,159877.OF,159880.OF,159881.OF,159883.OF,159885.OF,159886.OF,159887.OF,159888.OF,159889.OF,159890.OF,159891.OF,159895.OF,159896.OF,159897.OF,159898.OF,159899.OF,159901.OF,159902.OF,159903.OF,159905.OF,159906.OF,159907.OF,159908.OF,159909.OF,159910.OF,159912.OF,159913.OF,159915.OF,159916.OF,159918.OF,159919.OF,159922.OF,159923.OF,159925.OF,159928.OF,159929.OF,159930.OF,159931.OF,159932.OF,159933.OF,159935.OF,159936.OF,159938.OF,159939.OF,159940.OF,159943.OF,159944.OF,159945.OF,159948.OF,159949.OF,159951.OF,159952.OF,159955.OF,159956.OF,159957.OF,159958.OF,159959.OF,159960.OF,159961.OF,159962.OF,159964.OF,159965.OF,159966.OF,159967.OF,159968.OF,159969.OF,159970.OF,159971.OF,159973.OF,159974.OF,159975.OF,159976.OF,159977.OF,159978.OF,159979.OF,159982.OF,159983.OF,159984.OF,159986.OF,159987.OF,159990.OF,159991.OF,159992.OF,159993.OF,159994.OF,159995.OF,159996.OF,159997.OF,159998.OF,159999.OF,160119.OF,160127.OF,160133.OF,160135.OF,160137.OF,160144.OF,160218.OF,160219.OF,160221.OF,160222.OF,160223.OF,160224.OF,160225.OF,160322.OF,160418.OF,160419.OF,160420.OF,160422.OF,160424.OF,160516.OF,160517.OF,160615.OF,160616.OF,160620.OF,160625.OF,160626.OF,160628.OF,160629.OF,160630.OF,160631.OF,160632.OF,160633.OF,160634.OF,160635.OF,160636.OF,160637.OF,160638.OF,160639.OF,160643.OF,160646.OF,160706.OF,160716.OF,160724.OF,160725.OF,160806.OF,160807.OF,160814.OF,160919.OF,160925.OF,161017.OF,161022.OF,161024.OF,161025.OF,161026.OF,161027.OF,161028.OF,161029.OF,161030.OF,161031.OF,161032.OF,161033.OF,161035.OF,161036.OF,161037.OF,161039.OF,161118.OF,161121.OF,161122.OF,161123.OF,161211.OF,161217.OF,161227.OF,161604.OF,161607.OF,161612.OF,161613.OF,161628.OF,161631.OF,161715.OF,161718.OF,161720.OF,161721.OF,161723.OF,161724.OF,161725.OF,161726.OF,161811.OF,161812.OF,161816.OF,161907.OF,162107.OF,162208.OF,162213.OF,162216.OF,162307.OF,162412.OF,162413.OF,162416.OF,162509.OF,162510.OF,162711.OF,162714.OF,162907.OF,163109.OF,163110.OF,163111.OF,163113.OF,163114.OF,163115.OF,163116.OF,163118.OF,163209.OF,163407.OF,163808.OF,163821.OF,164205.OF,164304.OF,164401.OF,164402.OF,164508.OF,164809.OF,164811.OF,164818.OF,164825.OF,164905.OF,164908.OF,165309.OF,165310.OF,165312.OF,165511.OF,165515.OF,165519.OF,165520.OF,165521.OF,165522.OF,165523.OF,165524.OF,165525.OF,165806.OF,165810.OF,166402.OF,166802.OF,167301.OF,167302.OF,167503.OF,167506.OF,167507.OF,167601.OF,167702.OF,167703.OF,168203.OF,168204.OF,168205.OF,168701.OF,168702.OF,180003.OF,200002.OF,202015.OF,202017.OF,202021.OF,202025.OF,202211.OF,206012.OF,213010.OF,217016.OF,217017.OF,217019.OF,217027.OF,233010.OF,240014.OF,240016.OF,240019.OF,257060.OF,270010.OF,270026.OF,290010.OF,310318.OF,310398.OF,320010.OF,320014.OF,320020.OF,320022.OF,340006.OF,360001.OF,370023.OF,376510.OF,399001.OF,399011.OF,410008.OF,410010.OF,450008.OF,450009.OF,460220.OF,460300.OF,470007.OF,470068.OF,481009.OF,481012.OF,501005.OF,501006.OF,501007.OF,501008.OF,501009.OF,501010.OF,501011.OF,501012.OF,501016.OF,501019.OF,501025.OF,501029.OF,501030.OF,501031.OF,501036.OF,501037.OF,501043.OF,501045.OF,501047.OF,501048.OF,501050.OF,501057.OF,501058.OF,501059.OF,501060.OF,501061.OF,501067.OF,501068.OF,501069.OF,501086.OF,501089.OF,501090.OF,501219.OF,501301.OF,501302.OF,501303.OF,501305.OF,501306.OF,501307.OF,501308.OF,501309.OF,501310.OF,501311.OF,502000.OF,502003.OF,502006.OF,502010.OF,502013.OF,502020.OF,502023.OF,502040.OF,502048.OF,502053.OF,502056.OF,510010.OF,510020.OF,510030.OF,510050.OF,510060.OF,510090.OF,510100.OF,510110.OF,510120.OF,510130.OF,510150.OF,510160.OF,510170.OF,510180.OF,510190.OF,510200.OF,510210.OF,510220.OF,510230.OF,510270.OF,510290.OF,510300.OF,510310.OF,510330.OF,510350.OF,510360.OF,510370.OF,510380.OF,510390.OF,510410.OF,510440.OF,510500.OF,510510.OF,510530.OF,510550.OF,510560.OF,510570.OF,510580.OF,510590.OF,510600.OF,510630.OF,510650.OF,510660.OF,510680.OF,510710.OF,510760.OF,510770.OF,510800.OF,510810.OF,510850.OF,510880.OF,510990.OF,512000.OF,512010.OF,512040.OF,512070.OF,512090.OF,512100.OF,512120.OF,512150.OF,512160.OF,512170.OF,512180.OF,512190.OF,512200.OF,512220.OF,512260.OF,512280.OF,512290.OF,512320.OF,512330.OF,512360.OF,512380.OF,512390.OF,512400.OF,512480.OF,512500.OF,512510.OF,512520.OF,512530.OF,512550.OF,512560.OF,512570.OF,512580.OF,512590.OF,512600.OF,512610.OF,512640.OF,512650.OF,512660.OF,512670.OF,512680.OF,512690.OF,512700.OF,512710.OF,512720.OF,512730.OF,512750.OF,512760.OF,512770.OF,512780.OF,512800.OF,512810.OF,512820.OF,512870.OF,512880.OF,512890.OF,512900.OF,512910.OF,512920.OF,512930.OF,512950.OF,512960.OF,512970.OF,512980.OF,512990.OF,513020.OF,513070.OF,513090.OF,513150.OF,513160.OF,513200.OF,513230.OF,513320.OF,513550.OF,513590.OF,513660.OF,513680.OF,513690.OF,513700.OF,513770.OF,513860.OF,513900.OF,513960.OF,513980.OF,513990.OF,515000.OF,515010.OF,515020.OF,515030.OF,515050.OF,515060.OF,515070.OF,515080.OF,515090.OF,515100.OF,515110.OF,515120.OF,515130.OF,515150.OF,515160.OF,515170.OF,515180.OF,515190.OF,515200.OF,515210.OF,515220.OF,515230.OF,515250.OF,515260.OF,515280.OF,515290.OF,515300.OF,515310.OF,515320.OF,515330.OF,515350.OF,515360.OF,515380.OF,515390.OF,515400.OF,515450.OF,515500.OF,515510.OF,515520.OF,515530.OF,515550.OF,515560.OF,515570.OF,515580.OF,515590.OF,515600.OF,515610.OF,515630.OF,515650.OF,515660.OF,515670.OF,515680.OF,515700.OF,515710.OF,515750.OF,515760.OF,515770.OF,515780.OF,515790.OF,515800.OF,515810.OF,515820.OF,515850.OF,515860.OF,515870.OF,515880.OF,515890.OF,515900.OF,515910.OF,515920.OF,515930.OF,515950.OF,515960.OF,515980.OF,515990.OF,516000.OF,516010.OF,516020.OF,516050.OF,516060.OF,516070.OF,516080.OF,516090.OF,516100.OF,516110.OF,516120.OF,516130.OF,516150.OF,516160.OF,516180.OF,516190.OF,516200.OF,516210.OF,516220.OF,516260.OF,516270.OF,516290.OF,516300.OF,516310.OF,516320.OF,516330.OF,516350.OF,516360.OF,516380.OF,516390.OF,516480.OF,516500.OF,516510.OF,516520.OF,516530.OF,516550.OF,516560.OF,516570.OF,516580.OF,516590.OF,516600.OF,516610.OF,516620.OF,516630.OF,516640.OF,516650.OF,516660.OF,516670.OF,516680.OF,516690.OF,516700.OF,516710.OF,516720.OF,516730.OF,516750.OF,516760.OF,516770.OF,516780.OF,516790.OF,516800.OF,516810.OF,516820.OF,516830.OF,516850.OF,516860.OF,516880.OF,516890.OF,516900.OF,516910.OF,516920.OF,516930.OF,516950.OF,516960.OF,516970.OF,516980.OF,517000.OF,517010.OF,517030.OF,517050.OF,517080.OF,517090.OF,517100.OF,517110.OF,517120.OF,517160.OF,517170.OF,517180.OF,517200.OF,517270.OF,517280.OF,517300.OF,517330.OF,517350.OF,517360.OF,517380.OF,517390.OF,517500.OF,517550.OF,517660.OF,517760.OF,517770.OF,517780.OF,517800.OF,517880.OF,517900.OF,517960.OF,519027.OF,519032.OF,519034.OF,519100.OF,519116.OF,519117.OF,519180.OF,519193.OF,519300.OF,519606.OF,519671.OF,519673.OF,519677.OF,519686.OF,519706.OF,519714.OF,519935.OF,519965.OF,519975.OF,530010.OF,530015.OF,530018.OF,540006.OF,540007.OF,540008.OF,540009.OF,540010.OF,540012.OF,560000.OF,560050.OF,560100.OF,560500.OF,560560.OF,560600.OF,560660.OF,560800.OF,560880.OF,561100.OF,561120.OF,561130.OF,561300.OF,561310.OF,561350.OF,561500.OF,561550.OF,561600.OF,561800.OF,561900.OF,561910.OF,561990.OF,562300.OF,562360.OF,562500.OF,562510.OF,562520.OF,562800.OF,562880.OF,562900.OF,562910.OF,562950.OF,563000.OF,580008.OF,585001.OF,588000.OF,588050.OF,588060.OF,588080.OF,588090.OF,588150.OF,588180.OF,588280.OF,588300.OF,588310.OF,588330.OF,588360.OF,588380.OF,588390.OF,588400.OF,590007.OF,660008.OF,660011.OF,671030.OF,673090.OF,673100.OF,673101.OF,690008.OF,700002.OF,740101.OF,900029.OF,900030.OF,920002.OF,920003.OF,920922.OF,920923.OF,960000.OF,960022.OF,970041.OF,970042.OF,H161604.OF,H202017.OF','ths_fund_thscode_fund','')
    if data_fund.errorcode != 0:
        print('error:{}'.format(data_fund.errmsg))
    else:
        all_fund=THS_BD('000008.OF,000042.OF,000051.OF,000059.OF,000082.OF,000176.OF,000248.OF,000309.OF,000311.OF,000312.OF,000313.OF,000326.OF,000368.OF,000373.OF,000376.OF,000409.OF,000411.OF,000418.OF,000457.OF,000471.OF,000478.OF,000512.OF,000513.OF,000524.OF,000549.OF,000577.OF,000586.OF,000592.OF,000594.OF,000596.OF,000613.OF,000628.OF,000656.OF,000688.OF,000696.OF,000697.OF,000711.OF,000729.OF,000746.OF,000751.OF,000756.OF,000761.OF,000778.OF,000780.OF,000793.OF,000803.OF,000826.OF,000827.OF,000828.OF,000831.OF,000835.OF,000854.OF,000866.OF,000867.OF,000884.OF,000893.OF,000913.OF,000916.OF,000925.OF,000942.OF,000948.OF,000950.OF,000955.OF,000960.OF,000961.OF,000962.OF,000968.OF,000971.OF,000974.OF,000975.OF,000978.OF,000979.OF,000985.OF,000991.OF,000996.OF,001008.OF,001009.OF,001015.OF,001016.OF,001027.OF,001028.OF,001036.OF,001039.OF,001040.OF,001042.OF,001043.OF,001044.OF,001047.OF,001048.OF,001050.OF,001051.OF,001052.OF,001054.OF,001064.OF,001070.OF,001072.OF,001097.OF,001104.OF,001105.OF,001113.OF,001126.OF,001133.OF,001149.OF,001158.OF,001162.OF,001163.OF,001166.OF,001167.OF,001171.OF,001178.OF,001180.OF,001186.OF,001188.OF,001193.OF,001195.OF,001208.OF,001214.OF,001223.OF,001230.OF,001236.OF,001237.OF,001241.OF,001242.OF,001243.OF,001245.OF,001277.OF,001291.OF,001313.OF,001319.OF,001344.OF,001351.OF,001361.OF,001396.OF,001397.OF,001404.OF,001409.OF,001410.OF,001416.OF,001420.OF,001421.OF,001426.OF,001455.OF,001469.OF,001473.OF,001476.OF,001482.OF,001490.OF,001496.OF,001520.OF,001521.OF,001528.OF,001539.OF,001541.OF,001542.OF,001548.OF,001549.OF,001550.OF,001551.OF,001552.OF,001553.OF,001556.OF,001557.OF,001576.OF,001577.OF,001579.OF,001583.OF,001588.OF,001589.OF,001592.OF,001593.OF,001594.OF,001595.OF,001605.OF,001616.OF,001617.OF,001618.OF,001626.OF,001628.OF,001629.OF,001630.OF,001631.OF,001632.OF,001637.OF,001638.OF,001643.OF,001644.OF,001645.OF,001651.OF,001672.OF,001677.OF,001678.OF,001685.OF,001692.OF,001703.OF,001705.OF,001707.OF,001714.OF,001717.OF,001718.OF,001719.OF,001725.OF,001726.OF,001733.OF,001736.OF,001749.OF,001764.OF,001766.OF,001781.OF,001790.OF,001825.OF,001849.OF,001877.OF,001878.OF,001879.OF,001899.OF,001915.OF,001917.OF,001938.OF,001956.OF,001974.OF,001975.OF,002063.OF,002076.OF,002121.OF,002168.OF,002199.OF,002210.OF,002229.OF,002236.OF,002289.OF,002300.OF,002310.OF,002311.OF,002315.OF,002316.OF,002332.OF,002333.OF,002334.OF,002335.OF,002385.OF,002387.OF,002510.OF,002556.OF,002595.OF,002621.OF,002656.OF,002670.OF,002671.OF,002697.OF,002861.OF,002871.OF,002872.OF,002900.OF,002903.OF,002906.OF,002907.OF,002952.OF,002974.OF,002977.OF,002978.OF,002979.OF,002980.OF,002982.OF,002984.OF,002987.OF,003015.OF,003016.OF,003017.OF,003053.OF,003054.OF,003069.OF,003070.OF,003145.OF,003184.OF,003190.OF,003191.OF,003194.OF,003230.OF,003231.OF,003232.OF,003233.OF,003298.OF,003299.OF,003318.OF,003359.OF,003366.OF,003416.OF,003475.OF,003548.OF,003578.OF,003579.OF,003624.OF,003625.OF,003634.OF,003646.OF,003647.OF,003745.OF,003760.OF,003761.OF,003765.OF,003766.OF,003834.OF,003853.OF,003865.OF,003876.OF,003884.OF,003885.OF,003956.OF,003957.OF,003958.OF,003984.OF,003985.OF,003986.OF,004040.OF,004041.OF,004069.OF,004070.OF,004075.OF,004098.OF,004189.OF,004190.OF,004191.OF,004192.OF,004193.OF,004194.OF,004195.OF,004241.OF,004292.OF,004342.OF,004343.OF,004344.OF,004345.OF,004346.OF,004347.OF,004348.OF,004350.OF,004352.OF,004403.OF,004404.OF,004407.OF,004408.OF,004409.OF,004410.OF,004432.OF,004433.OF,004450.OF,004476.OF,004488.OF,004512.OF,004513.OF,004532.OF,004533.OF,004597.OF,004598.OF,004606.OF,004616.OF,004640.OF,004642.OF,004643.OF,004683.OF,004686.OF,004698.OF,004716.OF,004730.OF,004742.OF,004743.OF,004744.OF,004746.OF,004752.OF,004753.OF,004784.OF,004788.OF,004789.OF,004790.OF,004791.OF,004805.OF,004812.OF,004813.OF,004851.OF,004854.OF,004855.OF,004856.OF,004857.OF,004858.OF,004870.OF,004874.OF,004875.OF,004876.OF,004891.OF,004892.OF,004925.OF,004945.OF,004995.OF,004996.OF,004997.OF,005009.OF,005035.OF,005036.OF,005037.OF,005038.OF,005051.OF,005052.OF,005062.OF,005063.OF,005064.OF,005102.OF,005103.OF,005106.OF,005112.OF,005113.OF,005114.OF,005125.OF,005137.OF,005152.OF,005161.OF,005188.OF,005189.OF,005209.OF,005210.OF,005223.OF,005224.OF,005229.OF,005235.OF,005236.OF,005237.OF,005238.OF,005248.OF,005259.OF,005267.OF,005268.OF,005287.OF,005288.OF,005303.OF,005304.OF,005310.OF,005313.OF,005314.OF,005328.OF,005390.OF,005391.OF,005402.OF,005433.OF,005457.OF,005495.OF,005496.OF,005502.OF,005505.OF,005506.OF,005530.OF,005554.OF,005555.OF,005561.OF,005562.OF,005569.OF,005570.OF,005607.OF,005608.OF,005612.OF,005620.OF,005621.OF,005626.OF,005627.OF,005628.OF,005629.OF,005633.OF,005635.OF,005636.OF,005639.OF,005640.OF,005658.OF,005659.OF,005660.OF,005661.OF,005662.OF,005663.OF,005669.OF,005691.OF,005693.OF,005702.OF,005707.OF,005727.OF,005733.OF,005734.OF,005735.OF,005737.OF,005761.OF,005762.OF,005763.OF,005770.OF,005777.OF,005788.OF,005789.OF,005795.OF,005802.OF,005813.OF,005814.OF,005825.OF,005826.OF,005829.OF,005830.OF,005832.OF,005867.OF,005868.OF,005869.OF,005870.OF,005873.OF,005874.OF,005880.OF,005881.OF,005885.OF,005894.OF,005918.OF,005919.OF,005927.OF,005928.OF,005960.OF,005961.OF,005962.OF,005963.OF,005965.OF,005966.OF,005968.OF,005969.OF,005970.OF,005983.OF,005994.OF,005998.OF,005999.OF,006002.OF,006003.OF,006020.OF,006021.OF,006034.OF,006048.OF,006063.OF,006080.OF,006081.OF,006087.OF,006098.OF,006106.OF,006119.OF,006129.OF,006130.OF,006131.OF,006138.OF,006143.OF,006144.OF,006165.OF,006166.OF,006193.OF,006194.OF,006195.OF,006196.OF,006197.OF,006214.OF,006215.OF,006220.OF,006221.OF,006228.OF,006229.OF,006248.OF,006249.OF,006251.OF,006257.OF,006265.OF,006286.OF,006293.OF,006339.OF,006341.OF,006342.OF,006346.OF,006347.OF,006355.OF,006363.OF,006382.OF,006395.OF,006438.OF,006439.OF,006440.OF,006441.OF,006478.OF,006486.OF,006487.OF,006502.OF,006503.OF,006524.OF,006525.OF,006560.OF,006561.OF,006569.OF,006593.OF,006594.OF,006600.OF,006603.OF,006604.OF,006605.OF,006611.OF,006614.OF,006656.OF,006657.OF,006658.OF,006659.OF,006671.OF,006675.OF,006676.OF,006682.OF,006687.OF,006688.OF,006692.OF,006693.OF,006697.OF,006704.OF,006705.OF,006712.OF,006713.OF,006724.OF,006729.OF,006730.OF,006733.OF,006743.OF,006744.OF,006748.OF,006751.OF,006756.OF,006757.OF,006781.OF,006783.OF,006784.OF,006786.OF,006787.OF,006803.OF,006809.OF,006810.OF,006818.OF,006819.OF,006868.OF,006909.OF,006910.OF,006911.OF,006912.OF,006923.OF,006924.OF,006926.OF,006928.OF,006937.OF,006938.OF,006939.OF,006957.OF,006981.OF,007005.OF,007028.OF,007029.OF,007039.OF,007044.OF,007045.OF,007076.OF,007077.OF,007089.OF,007096.OF,007107.OF,007108.OF,007130.OF,007133.OF,007134.OF,007135.OF,007136.OF,007137.OF,007138.OF,007143.OF,007144.OF,007153.OF,007154.OF,007178.OF,007191.OF,007216.OF,007223.OF,007230.OF,007275.OF,007276.OF,007287.OF,007288.OF,007300.OF,007301.OF,007339.OF,007354.OF,007357.OF,007379.OF,007380.OF,007386.OF,007388.OF,007389.OF,007397.OF,007404.OF,007405.OF,007410.OF,007411.OF,007413.OF,007431.OF,007448.OF,007464.OF,007465.OF,007466.OF,007467.OF,007470.OF,007471.OF,007472.OF,007473.OF,007474.OF,007475.OF,007505.OF,007506.OF,007512.OF,007523.OF,007524.OF,007531.OF,007538.OF,007539.OF,007541.OF,007548.OF,007571.OF,007580.OF,007586.OF,007590.OF,007593.OF,007594.OF,007605.OF,007606.OF,007657.OF,007658.OF,007664.OF,007665.OF,007671.OF,007672.OF,007674.OF,007675.OF,007685.OF,007737.OF,007750.OF,007751.OF,007760.OF,007782.OF,007783.OF,007784.OF,007785.OF,007786.OF,007787.OF,007788.OF,007789.OF,007792.OF,007793.OF,007794.OF,007795.OF,007796.OF,007797.OF,007799.OF,007800.OF,007801.OF,007804.OF,007806.OF,007807.OF,007809.OF,007810.OF,007815.OF,007816.OF,007817.OF,007818.OF,007831.OF,007832.OF,007839.OF,007840.OF,007842.OF,007843.OF,007853.OF,007856.OF,007857.OF,007873.OF,007874.OF,007882.OF,007883.OF,007885.OF,007886.OF,007903.OF,007932.OF,007943.OF,007950.OF,007952.OF,007965.OF,007966.OF,007983.OF,007984.OF,007992.OF,007993.OF,007994.OF,007995.OF,008001.OF,008009.OF,008020.OF,008021.OF,008052.OF,008053.OF,008056.OF,008057.OF,008072.OF,008075.OF,008076.OF,008084.OF,008085.OF,008086.OF,008087.OF,008088.OF,008089.OF,008107.OF,008112.OF,008113.OF,008114.OF,008115.OF,008124.OF,008128.OF,008129.OF,008134.OF,008154.OF,008155.OF,008163.OF,008164.OF,008166.OF,008167.OF,008177.OF,008184.OF,008189.OF,008190.OF,008199.OF,008200.OF,008201.OF,008234.OF,008236.OF,008237.OF,008238.OF,008239.OF,008240.OF,008241.OF,008249.OF,008250.OF,008258.OF,008259.OF,008261.OF,008262.OF,008264.OF,008265.OF,008279.OF,008280.OF,008281.OF,008282.OF,008283.OF,008286.OF,008291.OF,008292.OF,008294.OF,008295.OF,008298.OF,008299.OF,008303.OF,008304.OF,008314.OF,008315.OF,008326.OF,008327.OF,008382.OF,008390.OF,008391.OF,008396.OF,008397.OF,008399.OF,008400.OF,008407.OF,008408.OF,008412.OF,008413.OF,008445.OF,008519.OF,008520.OF,008537.OF,008538.OF,008551.OF,008552.OF,008585.OF,008586.OF,008590.OF,008591.OF,008592.OF,008593.OF,008618.OF,008619.OF,008682.OF,008709.OF,008713.OF,008714.OF,008736.OF,008737.OF,008749.OF,008750.OF,008768.OF,008769.OF,008776.OF,008777.OF,008778.OF,008779.OF,008832.OF,008854.OF,008855.OF,008860.OF,008871.OF,008872.OF,008887.OF,008888.OF,008889.OF,008898.OF,008899.OF,008907.OF,008908.OF,008916.OF,008917.OF,008923.OF,008924.OF,008926.OF,008927.OF,008928.OF,008929.OF,008934.OF,008935.OF,008944.OF,008945.OF,008958.OF,008973.OF,008975.OF,008976.OF,009004.OF,009012.OF,009013.OF,009017.OF,009043.OF,009046.OF,009047.OF,009051.OF,009052.OF,009059.OF,009060.OF,009067.OF,009068.OF,009069.OF,009070.OF,009079.OF,009080.OF,009106.OF,009107.OF,009116.OF,009117.OF,009126.OF,009127.OF,009147.OF,009162.OF,009163.OF,009179.OF,009180.OF,009194.OF,009195.OF,009208.OF,009239.OF,009240.OF,009265.OF,009300.OF,009329.OF,009336.OF,009337.OF,009341.OF,009374.OF,009375.OF,009379.OF,009384.OF,009398.OF,009414.OF,009439.OF,009472.OF,009476.OF,009479.OF,009480.OF,009491.OF,009492.OF,009502.OF,009503.OF,009513.OF,009514.OF,009537.OF,009538.OF,009550.OF,009551.OF,009557.OF,009601.OF,009602.OF,009608.OF,009609.OF,009613.OF,009614.OF,009658.OF,009726.OF,009727.OF,009733.OF,009734.OF,009775.OF,009794.OF,009805.OF,009852.OF,009860.OF,009863.OF,009864.OF,009865.OF,009874.OF,009875.OF,009876.OF,009877.OF,009881.OF,009891.OF,009896.OF,009897.OF,009898.OF,009899.OF,009904.OF,009905.OF,009954.OF,009981.OF,009982.OF,009988.OF,010001.OF,010002.OF,010003.OF,010004.OF,010010.OF,010013.OF,010024.OF,010066.OF,010067.OF,010112.OF,010113.OF,010120.OF,010121.OF,010128.OF,010129.OF,010141.OF,010142.OF,010144.OF,010153.OF,010154.OF,010157.OF,010158.OF,010160.OF,010161.OF,010162.OF,010183.OF,010196.OF,010197.OF,010202.OF,010203.OF,010204.OF,010210.OF,010224.OF,010235.OF,010236.OF,010245.OF,010253.OF,010321.OF,010322.OF,010331.OF,010332.OF,010341.OF,010342.OF,010349.OF,010351.OF,010352.OF,010355.OF,010356.OF,010361.OF,010362.OF,010364.OF,010365.OF,010366.OF,010387.OF,010388.OF,010391.OF,010392.OF,010409.OF,010419.OF,010420.OF,010432.OF,010434.OF,010469.OF,010470.OF,010495.OF,010496.OF,010505.OF,010524.OF,010531.OF,010556.OF,010561.OF,010572.OF,010573.OF,010585.OF,010586.OF,010592.OF,010593.OF,010673.OF,010674.OF,010677.OF,010680.OF,010681.OF,010685.OF,010687.OF,010703.OF,010704.OF,010709.OF,010710.OF,010736.OF,010737.OF,010749.OF,010750.OF,010769.OF,010770.OF,010771.OF,010772.OF,010777.OF,010778.OF,010785.OF,010786.OF,010801.OF,010802.OF,010805.OF,010806.OF,010812.OF,010815.OF,010854.OF,010855.OF,010872.OF,010873.OF,010892.OF,010893.OF,010896.OF,010897.OF,010908.OF,010909.OF,010953.OF,010954.OF,010955.OF,010956.OF,010957.OF,010961.OF,010989.OF,010990.OF,010992.OF,010993.OF,010998.OF,011013.OF,011014.OF,011035.OF,011036.OF,011040.OF,011041.OF,011042.OF,011043.OF,011066.OF,011077.OF,011102.OF,011103.OF,011104.OF,011107.OF,011108.OF,011122.OF,011123.OF,011125.OF,011126.OF,011132.OF,011133.OF,011142.OF,011143.OF,011146.OF,011147.OF,011149.OF,011150.OF,011179.OF,011180.OF,011205.OF,011220.OF,011221.OF,011224.OF,011225.OF,011229.OF,011230.OF,011236.OF,011237.OF,011254.OF,011255.OF,011269.OF,011270.OF,011285.OF,011286.OF,011295.OF,011316.OF,011317.OF,011319.OF,011320.OF,011321.OF,011322.OF,011323.OF,011326.OF,011328.OF,011329.OF,011349.OF,011350.OF,011373.OF,011374.OF,011377.OF,011378.OF,011384.OF,011385.OF,011418.OF,011419.OF,011424.OF,011425.OF,011426.OF,011457.OF,011463.OF,011470.OF,011473.OF,011476.OF,011478.OF,011498.OF,011499.OF,011506.OF,011507.OF,011512.OF,011513.OF,011545.OF,011546.OF,011589.OF,011590.OF,011592.OF,011601.OF,011602.OF,011607.OF,011608.OF,011609.OF,011610.OF,011611.OF,011612.OF,011613.OF,011614.OF,011615.OF,011630.OF,011631.OF,011645.OF,011646.OF,011651.OF,011652.OF,011667.OF,011668.OF,011685.OF,011686.OF,011722.OF,011723.OF,011726.OF,011793.OF,011824.OF,011825.OF,011832.OF,011833.OF,011836.OF,011839.OF,011840.OF,011853.OF,011854.OF,011860.OF,011861.OF,011882.OF,011883.OF,011923.OF,011926.OF,011934.OF,011935.OF,011966.OF,011967.OF,011971.OF,011972.OF,012040.OF,012041.OF,012042.OF,012043.OF,012044.OF,012045.OF,012046.OF,012080.OF,012081.OF,012116.OF,012117.OF,012126.OF,012127.OF,012151.OF,012152.OF,012157.OF,012158.OF,012179.OF,012180.OF,012181.OF,012206.OF,012207.OF,012210.OF,012211.OF,012212.OF,012213.OF,012214.OF,012215.OF,012238.OF,012244.OF,012245.OF,012275.OF,012276.OF,012315.OF,012316.OF,012319.OF,012320.OF,012321.OF,012322.OF,012323.OF,012326.OF,012327.OF,012340.OF,012341.OF,012362.OF,012363.OF,012364.OF,012365.OF,012371.OF,012372.OF,012394.OF,012401.OF,012402.OF,012405.OF,012414.OF,012415.OF,012416.OF,012417.OF,012419.OF,012445.OF,012461.OF,012462.OF,012496.OF,012497.OF,012498.OF,012499.OF,012503.OF,012504.OF,012516.OF,012517.OF,012537.OF,012538.OF,012543.OF,012544.OF,012547.OF,012548.OF,012549.OF,012550.OF,012551.OF,012552.OF,012553.OF,012554.OF,012555.OF,012559.OF,012560.OF,012561.OF,012562.OF,012590.OF,012596.OF,012597.OF,012598.OF,012599.OF,012600.OF,012605.OF,012606.OF,012619.OF,012620.OF,012629.OF,012630.OF,012634.OF,012635.OF,012636.OF,012637.OF,012643.OF,012644.OF,012645.OF,012646.OF,012655.OF,012658.OF,012663.OF,012664.OF,012679.OF,012680.OF,012694.OF,012695.OF,012696.OF,012697.OF,012698.OF,012699.OF,012700.OF,012708.OF,012709.OF,012712.OF,012713.OF,012717.OF,012718.OF,012722.OF,012723.OF,012724.OF,012725.OF,012728.OF,012729.OF,012730.OF,012731.OF,012733.OF,012734.OF,012737.OF,012738.OF,012754.OF,012755.OF,012756.OF,012757.OF,012759.OF,012760.OF,012761.OF,012762.OF,012763.OF,012764.OF,012768.OF,012769.OF,012781.OF,012782.OF,012800.OF,012801.OF,012802.OF,012808.OF,012809.OF,012810.OF,012811.OF,012831.OF,012832.OF,012835.OF,012836.OF,012837.OF,012838.OF,012842.OF,012854.OF,012855.OF,012857.OF,012862.OF,012863.OF,012872.OF,012873.OF,012874.OF,012875.OF,012882.OF,012883.OF,012884.OF,012885.OF,012886.OF,012894.OF,012895.OF,012898.OF,012899.OF,012900.OF,012901.OF,012907.OF,012908.OF,012911.OF,012912.OF,012926.OF,012927.OF,012930.OF,012961.OF,012962.OF,012969.OF,012970.OF,012973.OF,012974.OF,012989.OF,012990.OF,013004.OF,013005.OF,013011.OF,013012.OF,013013.OF,013014.OF,013019.OF,013020.OF,013031.OF,013032.OF,013035.OF,013048.OF,013049.OF,013050.OF,013053.OF,013054.OF,013074.OF,013080.OF,013081.OF,013082.OF,013083.OF,013084.OF,013093.OF,013094.OF,013105.OF,013106.OF,013119.OF,013120.OF,013121.OF,013122.OF,013125.OF,013126.OF,013129.OF,013130.OF,013132.OF,013133.OF,013134.OF,013151.OF,013152.OF,013162.OF,013163.OF,013174.OF,013177.OF,013178.OF,013179.OF,013180.OF,013188.OF,013195.OF,013196.OF,013218.OF,013219.OF,013233.OF,013234.OF,013242.OF,013262.OF,013273.OF,013275.OF,013276.OF,013277.OF,013278.OF,013286.OF,013291.OF,013292.OF,013298.OF,013299.OF,013302.OF,013303.OF,013304.OF,013305.OF,013306.OF,013307.OF,013310.OF,013311.OF,013313.OF,013314.OF,013315.OF,013316.OF,013317.OF,013318.OF,013319.OF,013320.OF,013330.OF,013331.OF,013332.OF,013339.OF,013340.OF,013355.OF,013413.OF,013414.OF,013415.OF,013416.OF,013437.OF,013438.OF,013442.OF,013443.OF,013444.OF,013445.OF,013446.OF,013447.OF,013448.OF,013453.OF,013465.OF,013466.OF,013471.OF,013472.OF,013473.OF,013474.OF,013475.OF,013476.OF,013477.OF,013478.OF,013479.OF,013480.OF,013481.OF,013490.OF,013491.OF,013502.OF,013503.OF,013505.OF,013511.OF,013596.OF,013597.OF,013601.OF,013602.OF,013605.OF,013606.OF,013613.OF,013622.OF,013639.OF,013640.OF,013641.OF,013642.OF,013802.OF,013803.OF,013810.OF,013811.OF,013816.OF,013817.OF,013833.OF,013834.OF,013869.OF,013870.OF,013873.OF,013874.OF,013878.OF,013879.OF,013881.OF,013899.OF,013900.OF,013907.OF,013908.OF,013918.OF,013919.OF,013922.OF,013923.OF,013926.OF,013927.OF,013940.OF,013941.OF,013942.OF,013943.OF,014018.OF,014019.OF,014028.OF,014046.OF,014052.OF,014062.OF,014063.OF,014064.OF,014099.OF,014100.OF,014110.OF,014111.OF,014117.OF,014118.OF,014125.OF,014126.OF,014128.OF,014129.OF,014130.OF,014133.OF,014134.OF,014153.OF,014154.OF,014155.OF,014156.OF,014163.OF,014172.OF,014173.OF,014174.OF,014185.OF,014186.OF,014187.OF,014188.OF,014191.OF,014192.OF,014193.OF,014194.OF,014198.OF,014201.OF,014202.OF,014218.OF,014219.OF,014226.OF,014227.OF,014237.OF,014238.OF,014256.OF,014299.OF,014300.OF,014303.OF,014304.OF,014331.OF,014332.OF,014344.OF,014345.OF,014364.OF,014372.OF,014376.OF,014377.OF,014380.OF,014382.OF,014406.OF,014407.OF,014416.OF,014417.OF,014418.OF,014419.OF,014431.OF,014432.OF,014453.OF,014455.OF,014519.OF,014520.OF,014528.OF,014529.OF,014530.OF,014531.OF,014532.OF,014533.OF,014534.OF,014535.OF,014536.OF,014543.OF,014544.OF,014602.OF,014603.OF,014604.OF,014605.OF,014629.OF,014630.OF,014631.OF,014673.OF,014674.OF,014689.OF,014690.OF,014736.OF,014737.OF,014758.OF,014776.OF,014777.OF,014820.OF,014821.OF,014831.OF,014832.OF,014863.OF,014864.OF,014866.OF,014902.OF,014903.OF,014906.OF,014907.OF,014908.OF,014909.OF,014930.OF,014932.OF,014937.OF,014942.OF,014943.OF,014967.OF,014979.OF,014980.OF,014983.OF,014984.OF,014985.OF,014994.OF,014997.OF,015001.OF,015002.OF,015037.OF,015038.OF,015040.OF,015042.OF,015046.OF,015048.OF,015052.OF,015053.OF,015060.OF,015075.OF,015085.OF,015086.OF,015114.OF,015115.OF,015123.OF,015124.OF,015139.OF,015140.OF,015157.OF,015159.OF,015170.OF,015171.OF,015172.OF,015176.OF,015177.OF,015178.OF,015194.OF,015195.OF,015196.OF,015197.OF,015198.OF,015199.OF,015225.OF,015226.OF,015278.OF,015279.OF,015328.OF,015329.OF,015336.OF,015337.OF,015364.OF,015387.OF,015388.OF,015395.OF,015468.OF,015475.OF,015484.OF,020011.OF,020021.OF,040002.OF,040180.OF,040190.OF,050002.OF,050013.OF,050021.OF,050024.OF,070023.OF,070030.OF,070039.OF,090010.OF,090012.OF,100032.OF,100038.OF,100053.OF,110003.OF,110019.OF,110020.OF,110021.OF,110022.OF,110026.OF,110030.OF,159601.OF,159602.OF,159606.OF,159608.OF,159610.OF,159611.OF,159613.OF,159618.OF,159619.OF,159701.OF,159702.OF,159703.OF,159706.OF,159707.OF,159708.OF,159709.OF,159710.OF,159711.OF,159712.OF,159713.OF,159715.OF,159716.OF,159717.OF,159718.OF,159719.OF,159720.OF,159721.OF,159723.OF,159725.OF,159726.OF,159728.OF,159729.OF,159730.OF,159731.OF,159732.OF,159733.OF,159735.OF,159736.OF,159738.OF,159739.OF,159743.OF,159745.OF,159748.OF,159751.OF,159752.OF,159755.OF,159757.OF,159758.OF,159760.OF,159761.OF,159763.OF,159766.OF,159767.OF,159768.OF,159769.OF,159770.OF,159773.OF,159775.OF,159776.OF,159777.OF,159778.OF,159779.OF,159780.OF,159781.OF,159782.OF,159783.OF,159786.OF,159787.OF,159788.OF,159789.OF,159790.OF,159791.OF,159792.OF,159793.OF,159795.OF,159796.OF,159801.OF,159804.OF,159805.OF,159806.OF,159807.OF,159808.OF,159810.OF,159811.OF,159813.OF,159814.OF,159815.OF,159819.OF,159820.OF,159821.OF,159824.OF,159825.OF,159827.OF,159828.OF,159835.OF,159836.OF,159837.OF,159838.OF,159839.OF,159840.OF,159841.OF,159842.OF,159843.OF,159845.OF,159847.OF,159848.OF,159849.OF,159851.OF,159852.OF,159853.OF,159855.OF,159856.OF,159857.OF,159858.OF,159859.OF,159861.OF,159862.OF,159863.OF,159864.OF,159865.OF,159867.OF,159869.OF,159870.OF,159871.OF,159872.OF,159873.OF,159875.OF,159876.OF,159877.OF,159880.OF,159881.OF,159883.OF,159885.OF,159886.OF,159887.OF,159888.OF,159889.OF,159890.OF,159891.OF,159895.OF,159896.OF,159897.OF,159898.OF,159899.OF,159901.OF,159902.OF,159903.OF,159905.OF,159906.OF,159907.OF,159908.OF,159909.OF,159910.OF,159912.OF,159913.OF,159915.OF,159916.OF,159918.OF,159919.OF,159922.OF,159923.OF,159925.OF,159928.OF,159929.OF,159930.OF,159931.OF,159932.OF,159933.OF,159935.OF,159936.OF,159938.OF,159939.OF,159940.OF,159943.OF,159944.OF,159945.OF,159948.OF,159949.OF,159951.OF,159952.OF,159955.OF,159956.OF,159957.OF,159958.OF,159959.OF,159960.OF,159961.OF,159962.OF,159964.OF,159965.OF,159966.OF,159967.OF,159968.OF,159969.OF,159970.OF,159971.OF,159973.OF,159974.OF,159975.OF,159976.OF,159977.OF,159978.OF,159979.OF,159982.OF,159983.OF,159984.OF,159986.OF,159987.OF,159990.OF,159991.OF,159992.OF,159993.OF,159994.OF,159995.OF,159996.OF,159997.OF,159998.OF,159999.OF,160119.OF,160127.OF,160133.OF,160135.OF,160137.OF,160144.OF,160218.OF,160219.OF,160221.OF,160222.OF,160223.OF,160224.OF,160225.OF,160322.OF,160418.OF,160419.OF,160420.OF,160422.OF,160424.OF,160516.OF,160517.OF,160615.OF,160616.OF,160620.OF,160625.OF,160626.OF,160628.OF,160629.OF,160630.OF,160631.OF,160632.OF,160633.OF,160634.OF,160635.OF,160636.OF,160637.OF,160638.OF,160639.OF,160643.OF,160646.OF,160706.OF,160716.OF,160724.OF,160725.OF,160806.OF,160807.OF,160814.OF,160919.OF,160925.OF,161017.OF,161022.OF,161024.OF,161025.OF,161026.OF,161027.OF,161028.OF,161029.OF,161030.OF,161031.OF,161032.OF,161033.OF,161035.OF,161036.OF,161037.OF,161039.OF,161118.OF,161121.OF,161122.OF,161123.OF,161211.OF,161217.OF,161227.OF,161604.OF,161607.OF,161612.OF,161613.OF,161628.OF,161631.OF,161715.OF,161718.OF,161720.OF,161721.OF,161723.OF,161724.OF,161725.OF,161726.OF,161811.OF,161812.OF,161816.OF,161907.OF,162107.OF,162208.OF,162213.OF,162216.OF,162307.OF,162412.OF,162413.OF,162416.OF,162509.OF,162510.OF,162711.OF,162714.OF,162907.OF,163109.OF,163110.OF,163111.OF,163113.OF,163114.OF,163115.OF,163116.OF,163118.OF,163209.OF,163407.OF,163808.OF,163821.OF,164205.OF,164304.OF,164401.OF,164402.OF,164508.OF,164809.OF,164811.OF,164818.OF,164825.OF,164905.OF,164908.OF,165309.OF,165310.OF,165312.OF,165511.OF,165515.OF,165519.OF,165520.OF,165521.OF,165522.OF,165523.OF,165524.OF,165525.OF,165806.OF,165810.OF,166402.OF,166802.OF,167301.OF,167302.OF,167503.OF,167506.OF,167507.OF,167601.OF,167702.OF,167703.OF,168203.OF,168204.OF,168205.OF,168701.OF,168702.OF,180003.OF,200002.OF,202015.OF,202017.OF,202021.OF,202025.OF,202211.OF,206012.OF,213010.OF,217016.OF,217017.OF,217019.OF,217027.OF,233010.OF,240014.OF,240016.OF,240019.OF,257060.OF,270010.OF,270026.OF,290010.OF,310318.OF,310398.OF,320010.OF,320014.OF,320020.OF,320022.OF,340006.OF,360001.OF,370023.OF,376510.OF,399001.OF,399011.OF,410008.OF,410010.OF,450008.OF,450009.OF,460220.OF,460300.OF,470007.OF,470068.OF,481009.OF,481012.OF,501005.OF,501006.OF,501007.OF,501008.OF,501009.OF,501010.OF,501011.OF,501012.OF,501016.OF,501019.OF,501025.OF,501029.OF,501030.OF,501031.OF,501036.OF,501037.OF,501043.OF,501045.OF,501047.OF,501048.OF,501050.OF,501057.OF,501058.OF,501059.OF,501060.OF,501061.OF,501067.OF,501068.OF,501069.OF,501086.OF,501089.OF,501090.OF,501219.OF,501301.OF,501302.OF,501303.OF,501305.OF,501306.OF,501307.OF,501308.OF,501309.OF,501310.OF,501311.OF,502000.OF,502003.OF,502006.OF,502010.OF,502013.OF,502020.OF,502023.OF,502040.OF,502048.OF,502053.OF,502056.OF,510010.OF,510020.OF,510030.OF,510050.OF,510060.OF,510090.OF,510100.OF,510110.OF,510120.OF,510130.OF,510150.OF,510160.OF,510170.OF,510180.OF,510190.OF,510200.OF,510210.OF,510220.OF,510230.OF,510270.OF,510290.OF,510300.OF,510310.OF,510330.OF,510350.OF,510360.OF,510370.OF,510380.OF,510390.OF,510410.OF,510440.OF,510500.OF,510510.OF,510530.OF,510550.OF,510560.OF,510570.OF,510580.OF,510590.OF,510600.OF,510630.OF,510650.OF,510660.OF,510680.OF,510710.OF,510760.OF,510770.OF,510800.OF,510810.OF,510850.OF,510880.OF,510990.OF,512000.OF,512010.OF,512040.OF,512070.OF,512090.OF,512100.OF,512120.OF,512150.OF,512160.OF,512170.OF,512180.OF,512190.OF,512200.OF,512220.OF,512260.OF,512280.OF,512290.OF,512320.OF,512330.OF,512360.OF,512380.OF,512390.OF,512400.OF,512480.OF,512500.OF,512510.OF,512520.OF,512530.OF,512550.OF,512560.OF,512570.OF,512580.OF,512590.OF,512600.OF,512610.OF,512640.OF,512650.OF,512660.OF,512670.OF,512680.OF,512690.OF,512700.OF,512710.OF,512720.OF,512730.OF,512750.OF,512760.OF,512770.OF,512780.OF,512800.OF,512810.OF,512820.OF,512870.OF,512880.OF,512890.OF,512900.OF,512910.OF,512920.OF,512930.OF,512950.OF,512960.OF,512970.OF,512980.OF,512990.OF,513020.OF,513070.OF,513090.OF,513150.OF,513160.OF,513200.OF,513230.OF,513320.OF,513550.OF,513590.OF,513660.OF,513680.OF,513690.OF,513700.OF,513770.OF,513860.OF,513900.OF,513960.OF,513980.OF,513990.OF,515000.OF,515010.OF,515020.OF,515030.OF,515050.OF,515060.OF,515070.OF,515080.OF,515090.OF,515100.OF,515110.OF,515120.OF,515130.OF,515150.OF,515160.OF,515170.OF,515180.OF,515190.OF,515200.OF,515210.OF,515220.OF,515230.OF,515250.OF,515260.OF,515280.OF,515290.OF,515300.OF,515310.OF,515320.OF,515330.OF,515350.OF,515360.OF,515380.OF,515390.OF,515400.OF,515450.OF,515500.OF,515510.OF,515520.OF,515530.OF,515550.OF,515560.OF,515570.OF,515580.OF,515590.OF,515600.OF,515610.OF,515630.OF,515650.OF,515660.OF,515670.OF,515680.OF,515700.OF,515710.OF,515750.OF,515760.OF,515770.OF,515780.OF,515790.OF,515800.OF,515810.OF,515820.OF,515850.OF,515860.OF,515870.OF,515880.OF,515890.OF,515900.OF,515910.OF,515920.OF,515930.OF,515950.OF,515960.OF,515980.OF,515990.OF,516000.OF,516010.OF,516020.OF,516050.OF,516060.OF,516070.OF,516080.OF,516090.OF,516100.OF,516110.OF,516120.OF,516130.OF,516150.OF,516160.OF,516180.OF,516190.OF,516200.OF,516210.OF,516220.OF,516260.OF,516270.OF,516290.OF,516300.OF,516310.OF,516320.OF,516330.OF,516350.OF,516360.OF,516380.OF,516390.OF,516480.OF,516500.OF,516510.OF,516520.OF,516530.OF,516550.OF,516560.OF,516570.OF,516580.OF,516590.OF,516600.OF,516610.OF,516620.OF,516630.OF,516640.OF,516650.OF,516660.OF,516670.OF,516680.OF,516690.OF,516700.OF,516710.OF,516720.OF,516730.OF,516750.OF,516760.OF,516770.OF,516780.OF,516790.OF,516800.OF,516810.OF,516820.OF,516830.OF,516850.OF,516860.OF,516880.OF,516890.OF,516900.OF,516910.OF,516920.OF,516930.OF,516950.OF,516960.OF,516970.OF,516980.OF,517000.OF,517010.OF,517030.OF,517050.OF,517080.OF,517090.OF,517100.OF,517110.OF,517120.OF,517160.OF,517170.OF,517180.OF,517200.OF,517270.OF,517280.OF,517300.OF,517330.OF,517350.OF,517360.OF,517380.OF,517390.OF,517500.OF,517550.OF,517660.OF,517760.OF,517770.OF,517780.OF,517800.OF,517880.OF,517900.OF,517960.OF,519027.OF,519032.OF,519034.OF,519100.OF,519116.OF,519117.OF,519180.OF,519193.OF,519300.OF,519606.OF,519671.OF,519673.OF,519677.OF,519686.OF,519706.OF,519714.OF,519935.OF,519965.OF,519975.OF,530010.OF,530015.OF,530018.OF,540006.OF,540007.OF,540008.OF,540009.OF,540010.OF,540012.OF,560000.OF,560050.OF,560100.OF,560500.OF,560560.OF,560600.OF,560660.OF,560800.OF,560880.OF,561100.OF,561120.OF,561130.OF,561300.OF,561310.OF,561350.OF,561500.OF,561550.OF,561600.OF,561800.OF,561900.OF,561910.OF,561990.OF,562300.OF,562360.OF,562500.OF,562510.OF,562520.OF,562800.OF,562880.OF,562900.OF,562910.OF,562950.OF,563000.OF,580008.OF,585001.OF,588000.OF,588050.OF,588060.OF,588080.OF,588090.OF,588150.OF,588180.OF,588280.OF,588300.OF,588310.OF,588330.OF,588360.OF,588380.OF,588390.OF,588400.OF,590007.OF,660008.OF,660011.OF,671030.OF,673090.OF,673100.OF,673101.OF,690008.OF,700002.OF,740101.OF,900029.OF,900030.OF,920002.OF,920003.OF,920922.OF,920923.OF,960000.OF,960022.OF,970041.OF,970042.OF,H161604.OF,H202017.OF','ths_fund_scale_fund','2017-04-01')
        fund_list=all_fund.data.dropna()["thscode"].to_list()
        date_list=['2017-03-01','2018-03-01','2019-03-01','2020-03-01','2021-03-01','2022-03-01']
#         data_df=[]
        for date in date_list:
            data_result=THS_BD(fund_list,'ths_fund_scale_fund',date)
            if data_result.errorcode != 0:
                print('error:{}'.format(data_result.errmsg))
            else:
                data_df=data_result.data
                data_df.to_csv('{}.csv'.format(date))



df_1= datapool_hd()

df_2= datapool_bd()

THS_iFinDLogout()
