==> 2. calculate information covariate.ipynb <==

import pandas as pd
import numpy as np
import datetime 
import scipy.stats as ss

import warnings
warnings.filterwarnings('ignore')

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
# import pandas_datareader as pdr
import seaborn
from data_pre import *

import statsmodels.api as sm
from statsmodels.regression.rolling import RollingOLS

seaborn.set_style("darkgrid")
pd.plotting.register_matplotlib_converters()
%matplotlib inline
path="/Users/effiehan/Desktop/term4/swsc/fdr_ratios/"

# the resulted information covariate and alpha

fund= pd.read_csv(path+"date_scale/2017-03-01.csv",index_col = 0,parse_dates = True)
fund_list=fund["thscode"].values.tolist()



data_return=get_data_return(fund_list)
data_return.head()

scale=get_scale(fund_list)
scale.head()

risk_free=get_rf()
risk_free.head()

carhart=get_carhart()
carhart.head()

ff=get_ff()
ff.head()

fund_size=get_fund_size(scale,"2020/03")
fund_size=fund_size.T
fund_size.to_csv(path+'cov/fund_size.csv')
fund_size.head()


return_gap=get_retrun_gap(data_return,"2020")
return_gap=return_gap.T
return_gap.to_csv(path+'cov/return_gap.csv')
return_gap.head()

rsquare,alpha,vol,p=cal_carhart(data_return,carhart,risk_free,"2020/03")
rsquare=rsquare.T
rsquare.to_csv(path+'cov/rsquare.csv')
rsquare.head()

vol=vol.T
vol.to_csv(path+'cov/vol.csv')
vol.head()

alpha.to_csv(path+'cov/alpha.csv')
alpha=alpha.T
alpha.head()

p.to_csv(path+'cov/p_values.csv')
p=p.T

sharp= get_sharp(data_return, risk_free,"2020/04") ###no infor at march lag for one month 
sharp=sharp.T
sharp=sharp.rename(columns={"2020/04": "2020/03",}, errors="raise")
sharp.to_csv(path+'cov/sharp.csv')
sharp.head()

beta, treynor=cal_capm(data_return,ff,risk_free,"2020/03")
beta=beta.T
beta.to_csv(path+'cov/beta.csv')
beta.head()

treynor=treynor.T
treynor.to_csv(path+'cov/treynor.csv')
treynor.head()

fund_flow=get_fund_flow(scale,data_return,"2020")
fund_flow=fund_flow.T
fund_flow=fund_flow.rename(columns={"flow": "2020/03",}, errors="raise")
fund_flow.to_csv(path+'cov/fund_flow.csv')
fund_flow.head()

cov=[rsquare,fund_size,return_gap,fund_flow,sharp,treynor,beta,vol]
combo=[]
pos_alpha=alpha[alpha['2020/03'].values > 0]
for i in cov:
    i=i[i.index.isin(pos_alpha.index)]
    i['rank'] = i['2020/03'].rank(ascending=False)
    i=i[["rank"]]
    combo.append(i)    


combo=[alpha]+[p]+combo
combo=pd.concat(combo, axis=1)
combo.columns=["alpha",'p','rsquare','fund_size','return_gap','fund_flow','sharp','treynor','beta','vol']
combo=combo.drop(combo[combo.index=="sum"].index)
combo

combo=combo.dropna()
combo

combo.to_csv(path+'cov/combo.csv')
